% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_cleaning.R
\name{advanced_cleaning}
\alias{advanced_cleaning}
\alias{remove_cell_duplicates}
\alias{move_2closest_cell}
\title{Advanced occurrence data cleaning}
\usage{
advanced_cleaning(data, x, y, raster_layer, cell_duplicates = TRUE,
                  move_points_inside = FALSE, move_limit_distance = NULL,
                  verbose = TRUE)

remove_cell_duplicates(data, x, y,
                       raster_layer)

move_2closest_cell(data, x, y, raster_layer,
                   move_limit_distance, verbose = TRUE)
}
\arguments{
\item{data}{data.frame with occurrence records. Rows with NA values will be
omitted.}

\item{x}{(character) name of the column in \code{data}
containing longitude values.}

\item{y}{(character) name of the column in \code{data}
containing latitude values.}

\item{raster_layer}{a raster layer (object of class
\code{\link[terra]{SpatRaster}}).}

\item{cell_duplicates}{(logical) whether to remove duplicate coordinates
considering raster cells. Default = TRUE.}

\item{move_points_inside}{(logical) whether to move records outside of raster
cells with valid values to the closest cell with values. Default = FALSE.}

\item{move_limit_distance}{maximum distance to move records outside cells
with valid values. Default = NULL. Must be defined if \code{move_points_inside}
= TRUE.}

\item{verbose}{(logical) whether to print messages of progress. Default =
TRUE.}
}
\value{
A data.frame with occurrence records resulting from advanced cleaning
procedures. Other columns will be added to describe changes made in the
original data.
}
\description{
Advanced processes of data cleaning involving duplicate removal
and movement of records.
}
\details{
Data used in this functions should have gone through initial processes of
cleaning and filtering.
}
\seealso{
\code{\link{initial_cleaning}}
}
