% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{bias}
\alias{bias}
\title{Example Bias File}
\format{
A \code{SpatRaster} object.
}
\value{
No return value. Used with function \code{\link[terra]{rast}} to
bring raster variables to analysis.
}
\description{
A \code{SpatRaster} object representing a bias layer used for extracting
background points with the \code{prepare_data()} function.
}
\examples{
bias <- terra::rast(system.file("extdata", "bias_file.tif",
                                package = "kuenm2"))

terra::plot(bias)
}
