% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_response.R
\name{bivariate_response}
\alias{bivariate_response}
\title{Bivariate response plot for fitted models}
\usage{
bivariate_response(models, variable1 , variable2, modelID = NULL, n = 500,
                   new_data = NULL, extrapolate = TRUE, add_bar = TRUE ,
                   add_limits = TRUE, color_palette	= NULL,
                   xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{models}{an object of class \code{fitted_models} returned by the
\code{\link[enmpa:fit_glms]{enmpa::fit_selected()}} function.}

\item{variable1}{(character) name of the variable to be plotted in x axis.}

\item{variable2}{(character) name of the variable to be plotted in y axis.}

\item{modelID}{(character) name of the ModelID presents in the fitted
object. Default = NULL.}

\item{n}{(numeric) the number of breaks for plotting grid. Default = 500}

\item{new_data}{a \code{SpatRaster}, data.frame, or  matrix of variables
representing an area of interest. Default = NULL.}

\item{extrapolate}{(logical) whether to allow extrapolation to study the
behavior of the response outside the calibration limits. Ignored if
\code{new_data} is defined. Default = TRUE.}

\item{add_bar}{(logical) whether to add bar legend. Default = TRUE.}

\item{add_limits}{(logical) whether to add calibration limits if
\code{extrapolate = TRUE}. Default = TRUE.}

\item{color_palette}{(function) a color palette function to be used to assign
colors in the plot. The default, NULL uses \code{rev(hcl.colors(n, "terrain"))}.}

\item{xlab}{(character) a label for the x axis. The default, NULL, uses the
name defined in \code{variable1}.}

\item{ylab}{(character) a label for the y axis. The default, NULL, uses the
name defined in \code{variable2}.}

\item{...}{additional arguments passed to \code{\link[graphics]{image}}.}
}
\value{
A bivariate plot considering \code{variable1} and \code{variable2}.
}
\description{
A plot of suitability prediction in a two-dimensional environmental space.
}
\examples{
# Example with glmnet
# Import example of fitted_models (output of fit_selected())
data(fitted_model_maxnet, package = "kuenm2")

# Response curve (notice response affected by covariance)
bivariate_response(models = fitted_model_maxnet, modelID = "Model_219",
                   variable1 = "bio_1", variable2 = "bio_12")

# Example with glm
# Import example of fitted_models (output of fit_selected())
data(fitted_model_glm, package = "kuenm2")

# Response curve
bivariate_response(models = fitted_model_glm, modelID = "Model_85",
                   variable1 = "bio_1", variable2 = "bio_7")
}
\seealso{
\code{\link[enmpa:response_curve]{enmpa::response_curve()}}
}
