% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_glmnetmx.R
\name{calibration_glmnetmx}
\alias{calibration_glmnetmx}
\title{Calibration routines using Maxent-like glmnet models}
\usage{
calibration_glmnetmx(
  data,
  formula_grid,
  test_concave = TRUE,
  addsamplestobackground = TRUE,
  use_weights = FALSE,
  parallel = TRUE,
  ncores = 1,
  progress_bar = TRUE,
  write_summary = FALSE,
  out_dir = NULL,
  parallel_type = "doSNOW",
  return_replicate = TRUE,
  omrat_thr = c(5, 10),
  omrat_threshold = 10,
  AIC = "ws",
  delta_aic = 2,
  allow_tolerance = TRUE,
  tolerance = 0.01,
  skip_existing_models = FALSE,
  verbose = TRUE
)
}
\description{
Calibration routines using Maxent-like glmnet models
}
