% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_prepare_data.R
\name{calibration_grid_glm}
\alias{calibration_grid_glm}
\title{Calibration Grid Generation using GLM}
\usage{
calibration_grid_glm(
  occ_bg,
  min_number = 2,
  min_continuous = NULL,
  categorical_var = NULL,
  features = c("l", "q", "lq", "lqp", "p")
)
}
\description{
Calibration Grid Generation using GLM
}
