% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_grid_glmnetmx.R
\name{calibration_grid_glmnetmx}
\alias{calibration_grid_glmnetmx}
\title{Grid of parameters for calibration of Maxent-like glmnet models}
\usage{
calibration_grid_glmnetmx(
  var_names = NULL,
  swd = NULL,
  x = NULL,
  y = NULL,
  fold_column = "folds",
  min_number = 2,
  categorical_var = NULL,
  features = c("l", "q", "lq", "lqp", "p"),
  min_continuous = NULL,
  regm = c(0.1, 1, 2, 3, 5)
)
}
\description{
Grid of parameters for calibration of Maxent-like glmnet models
}
