% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{chelsa_current}
\alias{chelsa_current}
\title{SpatRaster Representing present-day Conditions (CHELSA)}
\format{
A \code{SpatRaster} object.
}
\value{
No return value. Used with function \code{\link[terra]{rast}} to
bring raster variables to analysis.
}
\description{
Raster layer containing bioclimatic variables representing present-day
climatic conditions. The variables were resampled to a 10 arc-minute
resolution and masked using the \code{m} region provided in the package. Data
sourced from CHELSA:
\url{https://chelsa-climate.org/}
}
\examples{
chelsa_current <- terra::rast(system.file("extdata",
                                           "Current_CHELSA.tif",
                                           package = "kuenm2"))
terra::plot(chelsa_current)
}
