% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{chelsa_lgm_miroc}
\alias{chelsa_lgm_miroc}
\title{SpatRaster Representing LGM Conditions (GCM: MIROC-ESM)}
\format{
A \code{SpatRaster} object.
}
\value{
No return value. Used with function \code{\link[terra]{rast}} to
bring raster variables to analysis.
}
\description{
Raster layer containing bioclimatic variables representing Last Glacial
Maximum (LGM) climatic conditions based on the MIROC-ESM General Circulation
Model. The variables were resampled to 10arc-minutes and masked using the \code{m}
provided in the package. Data sourced from CHELSA:
\url{https://chelsa-climate.org/last-glacial-maximum-climate/}
}
\examples{
chelsa_lgm_miroc <- terra::rast(system.file("extdata",
                                           "CHELSA_LGM_MIROC-ESM.tif",
                                           package = "kuenm2"))
terra::plot(chelsa_lgm_miroc)
}
