% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enmeval_block}
\alias{enmeval_block}
\title{Spatial Blocks from ENMeval}
\format{
A \code{list} with the following elements:
\describe{
\item{occs.grp}{A \code{numeric} vector indicating the spatial group to which each occurrence belongs}
\item{bg.grp}{A \code{numeric} vector indicating the spatial group to which each background point belongs}
}
}
\usage{
data("enmeval_block")
}
\description{
A list resulting from \code{ENMeval::get.block()} to partition occurrence and background localities into bins for training and validation (or, evaluation and calibration). This object is used in the "Prepare Data for Model Calibration" vignette to demonstrate how to implement custom data partitions generated by \code{ENMeval} in \code{kuenm2}.
}
\keyword{datasets}
