% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_calibration_hist.R
\name{explore_calibration_hist}
\alias{explore_calibration_hist}
\title{Explore variable distribution for occurrence and background points}
\usage{
explore_calibration_hist(data, color_background = "#0000FF80",
                             color_presence = "#FF000080",
                             mfrow = c(1, 1), plot_median = TRUE, ...)
}
\arguments{
\item{data}{an object of class \code{prepare_data} returned by the
\code{\link{prepare_data}}() function.}

\item{color_background}{(character) color used to fill the histogram bars for
background data. Default is "#0000FF80".}

\item{color_presence}{(character) color used to fill the histogram bars for
presence data. Default is "#FF000080".}

\item{mfrow}{(numeric) a vector specifying the number of rows and columns in
the plot layout, e.g., c(rows, columns). Default is c(1, 1), meaning
one plot per variable.}

\item{plot_median}{(logical) whether to plot a vertical dashed line
representing the median of the presence points on the histogram.
Default is TRUE.}

\item{...}{further arguments and graphical parameters passed to
\code{\link{hist}}(), allowing customization of titles, axes, and
other elements.}
}
\description{
This function generates overlaid histograms to visualize the distribution of
predictor variables for occurrence (presence) and background points.
}
\examples{
#Import raster layers
var <- terra::rast(system.file("extdata", "Current_variables.tif",
                               package = "kuenm2"))
#Import occurrences
data(occ_data, package = "kuenm2")
#Prepare data for glmnet model
sp_swd <- prepare_data(model_type = "glmnet", occ = occ_data,
                       species = occ_data[1, 1], x = "x", y = "y",
                       spat_variables = var, mask = NULL,
                       categorical_variables = "SoilType",
                       do_pca = FALSE, deviance_explained = 95,
                       min_explained = 5, center = TRUE, scale = TRUE,
                       write_pca = FALSE, output_pca = NULL, nbg = 2000,
                       kfolds = 4, weights = NULL, min_number = 2,
                       min_continuous = NULL,
                       features = c("l", "q", "p", "lq", "lqp"),
                       regm = c(0.1, 1, 2, 3, 5),
                       include_xy = TRUE,
                       write_file = FALSE, file_name = NULL,
                       seed = 1)
#Explore calibration data
explore_calibration_hist(data = sp_swd,
                         color_background = "#0000FF80",
                         color_presence = "#FF000080",
                         mfrow = c(2, 3), plot_median = TRUE,
                         breaks = "Scott")

}
