% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_partition_geo.R
\name{explore_partition_geo}
\alias{explore_partition_geo}
\title{Explore the spatial distribution of partitions for occurrence and background points}
\usage{
explore_partition_geo(data, raster_variables, mask = NULL,
                      show_partitions = TRUE, partition_palette = "cols25",
                      custom_partition_palette = NULL, pr_col = "#D55E00",
                      bg_col = "#0072B2", pr_bg_col = "#CC79A7",
                      calibration_area_col = "gray80", ...)
}
\arguments{
\item{data}{an object of class \code{prepared_data} returned by the
\code{\link[=prepare_data]{prepare_data()}} function.}

\item{raster_variables}{(SpatRaster) predictor variables used for model
calibration.}

\item{mask}{(SpatRaster, SpatVector, or SpatExtent) spatial object used to
mask \code{raster_variables} to the area where the model will be calibrated.
Preferably the same object used in \code{prepare_data} (if applicable). Default is
NULL.}

\item{show_partitions}{(logical) whether to return \code{SpatRaster} showing the
spatial distribution of each partition for presence and background points.
Default is TRUE.}

\item{partition_palette}{(character) the color palette used to color the
different partitions. See \code{?kuenm2_discrete_palettes} to check available
options. Default is \code{"cols25"}. Only applicable if \code{show_partitions = TRUE}.}

\item{custom_partition_palette}{(character) a character vector defining
custom colors for the different partitions. The number of values must match
the number of partitions in \code{data}. Default is NULL, meaning the palette
defined in \code{partition_palette} will be used.}

\item{pr_col}{(character) the color used for cells with presence records.
Default is "#D55E00".}

\item{bg_col}{(character) the color used for cells with background points.
Default is "#0072B2".}

\item{pr_bg_col}{(character) the color used for cells with presences and
background points. Default is "#CC79A7".}

\item{calibration_area_col}{(character) the color used for cells without
presences or background points. Default is "gray80".}

\item{...}{additional arguments passed to \code{terra::plot()}.}
}
\value{
A categorical \code{SpatRaster} with four factor values representing:
\describe{
\item{1 - Background cells}{}
\item{2 - Presence cells}{}
\item{3 - Cells with both presence and background}{}
\item{4 - Non-used cells}{}
}

If show_partitions = TRUE, it also returns \code{SpatRaster} showing the spatial
distribution of each partition for presence and background points.
}
\description{
Explore the spatial distribution of partitions for occurrence and background points
}
\examples{
# Import raster layers
var <- terra::rast(system.file("extdata", "Current_variables.tif",
                               package = "kuenm2"))

# Import prepared_data
data(sp_swd, package = "kuenm2")

Explore partitions in the geographic space
pbg <- explore_partition_geo(data = sp_swd, raster_variables = var[[1]])
terra::plot(pbg)
}
