% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_occurrence_variables.R
\name{extract_occurrence_variables}
\alias{extract_occurrence_variables}
\title{Extracts Environmental Variables for Occurrences}
\usage{
extract_occurrence_variables(occ, x, y, raster_variables)
}
\arguments{
\item{occ}{A data.frame containing occurrence data. It must include columns
with longitude (x) and latitude (y) coordinates.}

\item{x}{(character) a string specifying the name of the column in occ that
contains the longitude values.}

\item{y}{(character) a string specifying the name of the column in occ that
contains the latitude values.}

\item{raster_variables}{(SpatRaster) predictor variables used to calibrate
the models.}
}
\value{
A data.frame containing the original x and y coordinates of the occurrence
points (\code{x} and \code{y}), the values of the variables extracted
from \code{raster_variables}, and a new column \code{pr_bg} with a value of 1 for all
occurrences.
}
\description{
This function extracts values from environmental or predictor variables
(\code{SpatRaster}) for georeferenced occurrence points. It also adds a column
indicating that these are presence points(pr_bg = 1).
}
\examples{
# Import occurrences
data(occ_data, package = "kuenm2")

# Import raster layers
var <- terra::rast(system.file("extdata", "Current_variables.tif",
                               package = "kuenm2"))

# Extracts environmental variables for occurrences
occ_var <- extract_occurrence_variables(occ = occ_data, x = "x", y = "y",
                                        raster_variables = var)
}
