% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_var_from_formulas.R
\name{extract_var_from_formulas}
\alias{extract_var_from_formulas}
\title{Extract predictor names from formulas}
\usage{
extract_var_from_formulas(formulas, ...)
}
\arguments{
\item{formulas}{(character or formula) model formulas.}

\item{...}{Arguments to pass to \code{\link[=all.vars]{all.vars()}}}
}
\value{
A character vector or a list of the same length as \code{formulas},
containing the names of the predictors each formula.
}
\description{
Extract predictor names from formulas
}
\examples{
# Import an example of calibration results
data(calib_results_maxnet, package = "kuenm2")

# Extract predictor names
vars <- extract_var_from_formulas(calib_results_maxnet$formula_grid$Formulas)
}
