% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_selected.R
\name{fit_selected}
\alias{fit_selected}
\title{Fit models selected after calibration}
\usage{
fit_selected(calibration_results, n_replicates = 1, rep_type = "kfold",
             train_portion = 0.7, write_models = FALSE, file_name = NULL,
             parallel = FALSE, ncores = NULL, progress_bar = TRUE,
             verbose = TRUE, seed = 42)
}
\arguments{
\item{calibration_results}{an object of class \code{calibration_results} returned
by the \code{\link{calibration}}() function.}

\item{n_replicates}{(numeric) the number of model replicates. Using the
default, 1, implies that one replicate is fit with all the data.}

\item{rep_type}{(character) the replicate type. It can be: "kfold", "bootstrap",
or "subsample". Default is "kfold".}

\item{train_portion}{(numeric) the proportion of occurrence records used to
train the model in each replicate. This parameter is applicable only when
\code{rep_type} is set to "bootstrap" or "subsample". Default is 0.7.}

\item{write_models}{(logical) whether to save the final fitted models to disk.
Default is FALSE.}

\item{file_name}{(character) the file name, with or without a path, for saving
the final models. This is only applicable if \code{write_models = TRUE}.}

\item{parallel}{(logical) whether to fit the final models in parallel. Default
is FALSE.}

\item{ncores}{(numeric) number of cores to use for parallel processing.
Default is NULL and uses available cores - 1. This is only applicable if
\code{parallel = TRUE}.}

\item{progress_bar}{(logical) whether to display a progress bar during processing.
Default is TRUE.}

\item{verbose}{(logical) whether to display detailed messages during processing.
Default is TRUE.}

\item{seed}{(numeric) integer value used to specify an initial seed to split
the data. Default is 1.}
}
\value{
An object of class 'fitted_models' containing the following elements:
\item{species}{a character string with the name of the species.}
\item{Models}{a list of fitted models, including replicates (trained with
the parts of the data) and full models (trained with all available records).}
\item{calibration_data}{a data.frame containing a column (\code{pr_bg}) that
identifies occurrence points (1) and background points (0), along with the
corresponding values of predictor variables for each point.}
\item{selected_models}{a data frame with the ID and summary of evaluation
metrics for the selected models.}
\item{weights}{a numeric vector specifying weights for the predictor variables
(if used).}
\item{pca}{a list of class \code{\link[terra]{prcomp}} representing the
result of principal component analysis (if performed).}
\item{addsamplestobackground}{a logical value indicating whether any presence
sample not already in the background was added.}
\item{omission_rate}{the omission rate determined during the calibration step.}
\item{thresholds}{the thresholds to binarize each replicate and the consensus
(mean and median), calculated based on the omission rate set in
\code{\link{calibration}}()}
}
\description{
This function fits models selected after candidate model training and testing
using the function \code{\link{calibration}}().
}
\details{
This function also computes model consensus (mean and median), the thresholds
to binarize model predictions based on the omission rate set during model
calibration to select models.
}
\examples{
# An example with maxnet models
data(calib_results_maxnet, package = "kuenm2")

# Fit models using calibration results
fm <- fit_selected(calibration_results = calib_results_maxnet,
                   n_replicates = 2)

# Output the fitted models
fm

# An example with GLMs
data(calib_results_glm, package = "kuenm2")

# Fit models using calibration results
fm_glm <- fit_selected(calibration_results = calib_results_glm,
                       n_replicates = 2)

# Output the fitted models
fm_glm
}
