% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fitted_model_concave}
\alias{fitted_model_concave}
\title{Fitted model with concave curves}
\format{
A \code{fitted_models} with the following elements:
\describe{
\item{species}{Species names}
\item{Models}{A \code{list} with the fitted maxnet models (replicates and full models)}
\item{calibration_data}{A \code{data.frame} containing the variables extracted for presence and background points}
\item{continuous_variables}{A \code{character} indicating the names of the continuous variables}
\item{categorical_variables}{A \code{character} indicating the names of the categorical variables}
\item{selected_models}{A \code{data.frame} with formulas and evaluation metrics for each selected model}
\item{weights}{A \code{numeric} vector specifying weights for the occurrence records. \code{NULL} if no weights were set.}
\item{pca}{A \code{prcomp} object containing PCA results. \code{NULL} if PCA was not performed.}
\item{addsamplestobackground}{A \code{logical} value indicating whether to add any presence point not already included to the background.}
\item{omission_rate}{A \code{numeric} value indicating the omission rate used to evaluate models.}
\item{thresholds}{A \code{numeric} vector with thresholds used to binarize each replicate and the consensus (mean and median), calculated based on the omission rate defined in \code{calibration()}.}
\item{algorithm}{A \code{character} string indicating the algorithm used (maxnet).}
\item{partition_method}{A \code{character} string indicating the partitioning method used.}
\item{n_replicates}{A \code{numeric} value indicating the number of replicates or folds.}
}
}
\usage{
data("fitted_model_concave")
}
\description{
A maxnet \code{fitted_models} object resulting from \code{fit_selected()} with a model presenting concave curves.
}
\keyword{datasets}
