% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flexsdm_block}
\alias{flexsdm_block}
\title{Spatial Blocks from flexsdm}
\format{
A \code{list} with the following elements:
\describe{
\item{part}{A \code{tibble} object with information used in 'data' arguments and a additional column .part with partition group.}
\item{best_part_info}{A \code{tibble} with information about the best partition.}
}
}
\usage{
data("enmeval_block")
}
\description{
A list resulting from \code{flexsdm::part_sblock()}, used to partition occurrence and background localities into bins for training and evaluation. This object is used in the "Prepare Data for Model Calibration" vignette to demonstrate how to implement custom data partitions generated by \code{flexsdm} in \code{kuenm2}
}
\keyword{datasets}
