% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{future_2100_ssp585_miroc}
\alias{future_2100_ssp585_miroc}
\title{SpatRaster Representing Future Conditions (2081-2100, SSP585, GCM: MIROC6)}
\format{
A \code{SpatRaster} object.
}
\value{
No return value. Used with function \code{\link[terra]{rast}} to
bring raster variables to analysis.
}
\description{
A raster layer containing bioclimatic variables representing future climatic
conditions (2081-2100) based on the MIROC6 General Circulation Model
under the SSP585 scenario. The variables were obtained at a 10 arc-minute
resolution and masked using the \code{m} region provided in the package. Data
sourced from WorldClim: \url{https://worldclim.org/data/cmip6/cmip6climate.html}
}
\examples{
future_2100_ssp585_miroc <- terra::rast(system.file("extdata",
                                    "wc2.1_10m_bioc_MIROC6_ssp585_2081-2100.tif",
                                     package = "kuenm2"))
terra::plot(future_2100_ssp585_miroc)
}
