% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_mx.R
\name{glm_mx}
\alias{glm_mx}
\title{Maxent-like GLM models}
\usage{
glm_mx(formula, family = binomial(link = "cloglog"), data, weights = NULL, ...)
}
\description{
for more details see Fithian and Hastie (2013) DOI: 10.1214/13-AOAS667
}
