% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet_mx.R
\name{glmnet_mx}
\alias{glmnet_mx}
\title{Maxent-like glmnet models}
\usage{
glmnet_mx(
  p,
  data,
  f,
  regmult = 1,
  regfun = maxnet.default.regularization,
  addsamplestobackground = TRUE,
  weights = NULL,
  calculate_AIC = FALSE,
  ...
)
}
\description{
Maxent-like glmnet models
}
