% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_cleaning.R
\name{initial_cleaning}
\alias{initial_cleaning}
\alias{sort_columns}
\alias{remove_missing}
\alias{remove_duplicates}
\alias{remove_corrdinates_00}
\alias{filter_decimal_precision}
\title{Initial occurrence data cleaning steps}
\usage{
initial_cleaning(data, species_column, longitude_column, latitude_column,
                 other_columns = NULL, keep_all_columns = TRUE,
                 sort_columns = TRUE, remove_na = TRUE, remove_empty = TRUE,
                 remove_duplicates = TRUE, by_decimal_precision = TRUE,
                 decimal_precision = 0, longitude_precision = NULL,
                 latitude_precision = NULL)

sort_columns(data, species_column, longitude_column,
             latitude_column, keep_all_columns = FALSE)

remove_missing(data, columns = NULL, remove_na = TRUE,
               remove_empty = TRUE, keep_all_columns = TRUE)

remove_duplicates(data, columns = NULL, keep_all_columns = TRUE)

remove_corrdinates_00(data, longitude_column, latitude_column)

filter_decimal_precision(data, longitude_column,
                         latitude_column, decimal_precision = 0,
                         longitude_precision = NULL,
                         latitude_precision = NULL)
}
\arguments{
\item{data}{data.frame with occurrence records.}

\item{species_column}{(character) name of the column in \code{data}
containing species name.}

\item{longitude_column}{(character) name of the column in \code{data}
containing longitude values.}

\item{latitude_column}{(character) name of the column in \code{data}
containing latitude values.}

\item{other_columns}{(character) vector of other column name(s) in
\code{data} to be considered while performing cleaning steps, default = NULL.}

\item{keep_all_columns}{(logical) whether to keep all columns in \code{data}.
Default = TRUE.}

\item{sort_columns}{(logical) whether to sort species, longitude, and
latitude columns in \code{data}. Default = TRUE.}

\item{remove_na}{(logical) whether to remove NA values in the columns
considered. Default = TRUE.}

\item{remove_empty}{(logical) whether to remove empty (missing) values in
the columns considered. Default = TRUE.}

\item{remove_duplicates}{(logical) whether to remove duplicates in
the columns considered. Default = TRUE.}

\item{by_decimal_precision}{(logical) whether to remove certain records with
coordinate precision lower than that of the following three parameters.
Default = TRUE.}

\item{decimal_precision}{(numeric) decimal precision threshold for
coordinates. Default = 0. Ignored if the following two parameters are defined.}

\item{longitude_precision}{(numeric) decimal precision threshold for
longitude. Default = NULL.}

\item{latitude_precision}{(numeric) decimal precision threshold for
latitude. Default = NULL.}

\item{columns}{(character) vector of additional column name(s) in
\code{data} to be considered while removing missing or duplicate records,
default = NULL.}
}
\value{
A data.frame with resulting occurrence records.
}
\description{
Simple occurrence data cleaning procedures.
}
\details{
Function \code{initial_cleaning} helps to perform all simple steps of data
cleaning.
}
\seealso{
\code{\link{advanced_cleaning}}
}
