% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife_draft.R
\name{jackknife}
\alias{jackknife}
\title{Jackknife}
\usage{
jackknife(fitted, modelID, cv = 5)
}
\arguments{
\item{fitted}{object.}

\item{modelID}{(character) name of  ModelID from the fitted object.}

\item{cv}{(numeric) number of k-folds froms cross validation. Default = 5.}
}
\value{
list.
}
\description{
Jackknife
}
