% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuenm2.R
\docType{package}
\name{kuenm2-package}
\alias{kuenm2}
\alias{kuenm2-package}
\title{kuenm2: Detailed Development of Ecological Niche Models}
\description{
kuenm2 helps to perform critical phases of the ecological niche modeling
process. Data preparation analyses and explorations can be done to prepare
data. Model calibration (model selection) can be done by creating and testing
several candidate models. Handy options for producing final models,
evaluating such models, and assessing extrapolation risks are also included.
Tools for post-modeling analyses are implemented to allow further
exploration of results.
}
\author{
\strong{Maintainer}: Marlon E. Cobos \email{manubio13@gmail.com} (\href{https://orcid.org/0000-0002-2611-1767}{ORCID})

Authors:
\itemize{
  \item Weverton Trindade \email{wevertonf1993@gmail.com} (\href{https://orcid.org/0000-0003-2045-4555}{ORCID})
  \item Luis F. Arias-Giraldo \email{lfarias.giraldo@gmail.com} (\href{https://orcid.org/0000-0003-4861-8064}{ORCID})
  \item Luis Osorio-Olvera \email{luismurao@gmail.com} (\href{https://orcid.org/0000-0003-0701-5398}{ORCID})
  \item A. Townsend Peterson \email{town@ku.edu} (\href{https://orcid.org/0000-0003-0243-2379}{ORCID})
}

}
