% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kuenm2_discrete_palletes}
\alias{kuenm2_discrete_palletes}
\title{Discrete palettes based on pals R package}
\format{
A \code{list} with the following color palettes: "alphabet",
"alphabet2", "cols25", "glasbey", "kelly", "polychrome", "stepped",
"stepped2", "stepped3", "okabe", "tableau20", "tol", "tol.groundcover",
"trubetskoy", and "watlington"
}
\usage{
data("kuenm2_discrete_palletes")
}
\description{
Color palettes designed for discrete, categorical data. Palettes retrived from pals R package
}
\references{
Wright K (2023). pals: Color Palettes, Colormaps, and Tools to Evaluate
Them_. R package version 1.8, \url{https://CRAN.R-project.org/package=pals}.
}
\keyword{datasets}
