% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{m}
\alias{m}
\title{SpatVector Representing Calibration Area for \emph{Myrcia hatschbachii}}
\format{
A \code{Spatvector} object.
}
\value{
No return value. Used with function \code{\link[terra]{vect}} to
bring raster variables to analysis.
}
\description{
A spatial vector defining the calibration area used to extract background
points for fitting models of \emph{Myrcia hatschbachii}. The area was generated by
creating a minimum convex polygon around presence records (\code{occ_data}), then
applying a 300 km buffer.
}
\examples{
m <- terra::vect(system.file("extdata",
                             "m.gpkg",
                              package = "kuenm2"))
terra::plot(m)
}
