% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{model_calibration}
\alias{model_calibration}
\title{Calibration Function for Model Selection and Evaluation}
\usage{
model_calibration(
  data,
  formula_grid,
  test_concave = TRUE,
  addsamplestobackground = TRUE,
  use_weights = FALSE,
  parallel = TRUE,
  ncores = 1,
  progress_bar = TRUE,
  write_summary = FALSE,
  out_dir = NULL,
  parallel_type = "doSNOW",
  return_replicate = TRUE,
  omrat_thr = c(5, 10),
  omrat_threshold = 10,
  AIC = "ws",
  delta_aic = 2,
  allow_tolerance = TRUE,
  tolerance = 0.01,
  skip_existing_models = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{A dataset of class \code{prepare_data} containing the data to be used in model fitting.}

\item{formula_grid}{A grid of formulas of class \code{calibration_grid_glm} or \code{calibration_grid_glmnet}.}

\item{test_concave}{Logical. Should concave curves in quadratic models be tested? Default = TRUE.}

\item{addsamplestobackground}{Logical. Should samples be added to the background? Default = TRUE.}

\item{use_weights}{Logical. Should weights be used in the models? Default = FALSE.}

\item{parallel}{Logical. Should parallel processing be used? Default = TRUE.}

\item{ncores}{Integer. Number of cores to use for parallel processing. Default = 1.}

\item{progress_bar}{Logical. Should a progress bar be shown during parallel processing? Only works if \code{parallel_type} is "doSNOW". Default = TRUE.}

\item{write_summary}{Logical. Should a summary of each candidate evaluation be written to a file? Default = FALSE.}

\item{out_dir}{Character. The directory where summaries will be written if \code{write_summary} is TRUE. Default = NULL.}

\item{parallel_type}{Character. The type of parallel processing to use. Default = "doSNOW".}

\item{return_replicate}{Logical. Should replicate models be returned? Default = TRUE.}

\item{omrat_thr}{Numeric vector. Thresholds for omission rate. Default = c(5, 10).}

\item{omrat_threshold}{Numeric. Specific threshold for omission rate. Default = 10.}

\item{AIC}{Character. AIC criteria to use. Default = "ws".}

\item{delta_aic}{Numeric. The delta AIC threshold. Default = 2.}

\item{allow_tolerance}{Logical. If omission rate is higher than the set threshold, should the model with the minimum omission rate be selected? Default = TRUE.}

\item{tolerance}{Numeric. Tolerance level for model selection. Default = 0.01.}

\item{skip_existing_models}{Logical. Should existing models be skipped? Only works if \code{write_summary} is TRUE. Default = FALSE.}

\item{verbose}{Logical. Should detailed output be printed? Default = TRUE.}
}
\value{
A list or model object based on the specified model and parameters.
}
\description{
This function performs model calibration using a specified grid of formulas
and data of class \code{prepare_data}. It supports both Generalized Linear Models (GLM)
and GLMNET models, depending on the class of the \code{formula_grid} argument.
}
