% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modvar.R
\name{modvar}
\alias{modvar}
\title{Title}
\usage{
modvar(
  projection_file,
  by_replicate = TRUE,
  by_gcm = TRUE,
  by_model = TRUE,
  consensus = "mean",
  write_files = TRUE,
  output_dir = NULL,
  return_rasters = FALSE,
  progress_bar = FALSE,
  verbose = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{overwrite}{}
}
\description{
Title
}
