% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mop.R
\name{mop}
\alias{mop}
\alias{out_m}
\alias{mop_distance}
\title{Analysis of extrapolation risks using the MOP metric}
\usage{
mop(m, g, mop_type = "basic", distance = "euclidean", scale = FALSE,
    center = FALSE, fix_NA = FALSE, percent = 5, comp_each = 2000,
    parallel = FALSE, n_cores = NULL, progress_bar = TRUE)

out_m(m_matrix, g_matrix, type = c("basic", "simple", "detailed"))

mop_distance(m_matrix, g_matrix, distance = "euclidean", percent = 5,
             comp_each = 2000, parallel = FALSE, n_cores = NULL,
             progress_bar = TRUE)
}
\arguments{
\item{m}{SpatRaster or matrix of variables representing the original area
of interest (where a model was calibrated). If a matrix is used, each column
represents a variable.}

\item{g}{a SpatRaster or matrix of variables representing the area or
scenario where dissimilarity and non-analogous conditions will be detected
(where a model is projected). Variable names must match between \code{m} and
\code{g}.}

\item{mop_type}{(character) type of MOP analyses to be performed. Options
are: "basic", "simple", and "detailed".}

\item{distance}{(character) one of the options: "euclidean" or "mahalanobis".
Default = "euclidean".}

\item{scale}{scaling options (logical or numeric-alike) as in
\code{\link{scale}}. Default = FALSE.}

\item{center}{(logic or numeric-alike) center options as in
\code{\link{scale}}. Default = FALSE.}

\item{fix_NA}{(logical) whether to fix the layers so all cells with NA values
coincide among all layers.}

\item{percent}{(numeric) percentage of points of m (the closest ones) used to
derive mean environmental distances to each g point.}

\item{comp_each}{(numeric) number of points of the g matrix to be used for
distance calculations at a time (default = 2000). Increasing this number
requires more RAM.}

\item{rescale_mop}{(logical) whether or not to re-scale distances 0-1.
Default = FALSE. Re-scaling complicates comparisons of dissimilarity values
obtained from exercises in which the percentage is changed.}

\item{parallel}{(logical) if TRUE, calculations will be performed in parallel
using \code{n_cores} of the computer. Using this option will speed up the
analysis  but will demand more RAM. Default = FALSE.}

\item{n_cores}{(numeric) number of cores to be used in parallel processing.
Default = NULL, in which case all CPU cores on current host - 1 will be used.}

\item{progress_bar}{(logical) whether to show a progress bar for calculations,
default = TRUE. Valid when calculation are not run in parallel.}

\item{m_matrix}{matrix of variables representing the original area of
interest. Each column represents a variable.}

\item{g_matrix}{matrix of variables representing the area of model projection
(where dissimilarities and/or non-analogous conditions are to be detected).
Each column represents a variable. Variable names must match with those in
\code{m_matrix}}
}
\value{
A list of results containing:
\itemize{
\item summary.- a list with details on the data used in the analysis.
\item mop_basic.- a SpatRaster or matrix for the projection area (\code{g}) with
dissimilarity values, where the highest value represent conditions in which
at least one of the variables in the projection area are non-analogous to
those in the calibration area.
\item mop_simple.- a SpatRaster or matrix for the projection area (\code{g}) with
values representing how many variables in the projection area are
non-analogous to those in the calibration area.
\item mop_detailed.- a list containing:
\itemize{
\item interpretation_combined.- a data.frame to help identify the variables
that are non-analogous to \code{m} in the layers towards_low_combined and
towards_high_combined.
\item towards_low_end.- a SpatRaster or matrix for all variables representing
where non-analogous conditions were found towards low values of the
variable.
\item towards_high_end.- a SpatRaster or matrix for all variables representing
where non-analogous conditions were found towards high values of the
variable.
\item towards_low_combined.- a SpatRaster or vector with values representing
the identity of the variables found to have non-analogous conditions
towards low values. Interpretation requires the use of the table
interpretation_combined.
\item towards_high_combined.- a SpatRaster or vector with values representing
the identity of the variables found to have non-analogous conditions
towards high values. Interpretation requires the use of the table
interpretation_combined.
}
}
}
\description{
Analysis to calculate the mobility-oriented parity metric and
other sub-products to represent conditions in a projection area that are
non-analogous to those in a calibration area.
}
\details{
The options for the argument \code{mop_type} return results that differ in
the detail of how non-analogous conditions are identified. The option "basic"
makes calculation as proposed by Owens et al. (2013;
\url{https://doi.org/10.1016/j.ecolmodel.2013.04.011}). Other options involve
further analyses that help to identify non-analogous conditions in more
detail (see description of results returned).

When the variables used to represent environmental conditions in the areas of
interest differ considerably in their units, scaling and centering is
recommended. This is only valid when Euclidean distances are used.
}
