% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{new_fitted_models}
\alias{new_fitted_models}
\title{fitted_models Class Constructor}
\usage{
new_fitted_models(
  species,
  Models,
  calibration_data,
  selected_models,
  weights,
  pca,
  addsamplestobackground,
  omission_rate,
  thresholds
)
}
\description{
fitted_models Class Constructor
}
