% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{new_occ}
\alias{new_occ}
\title{Independent Species Occurrence}
\format{
A \code{data.frame} with the following columns:
\describe{
\item{species}{The species name.}
\item{x}{Longitude.}
\item{y}{Latitude.}
}
}
\usage{
data("new_occ")
}
\description{
A \code{data.frame} containing the coordinates of 82 occurrences of \emph{Myrcia hatschbachii} (a tree endemic to southern Brazil). The valid occurrences were
sourced from NeotropicTree (Oliveira-Filho, 2017) and were used as
independent data to test the models fitted with the \code{occ_data}.
}
\references{
Oliveira_Filho, A.T. 2017. NeoTropTree, Flora arbórea da Região Neotropical: Um banco de dados envolvendo biogeografia, diversidade e conservação. Universidade Federal de Minas Gerais. (http://www.neotroptree,info).
}
\keyword{datasets}
