% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_prepare_data.R
\name{new_prepare_data}
\alias{new_prepare_data}
\title{prepare_data Class Constructor}
\usage{
new_prepare_data(
  species,
  calibration_data,
  formula_grid,
  kfolds,
  data_xy,
  continuous_variables,
  categorical_variables,
  weights,
  pca,
  model_type
)
}
\description{
prepare_data Class Constructor
}
