% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{new_projection_data}
\alias{new_projection_data}
\title{prepared_proj Class Constructor}
\usage{
new_projection_data(
  res_present,
  res_past,
  res_future,
  raster_pattern,
  variables,
  pca
)
}
\description{
prepared_proj Class Constructor
}
