% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{occ_data}
\alias{occ_data}
\title{Species Occurrence}
\format{
A \code{data.frame} with the following columns:
\describe{
\item{species}{The species name.}
\item{x}{Longitude.}
\item{y}{Latitude.}
}
}
\usage{
data("occ_data")
}
\description{
A \code{data.frame} containing the coordinates of 51 valid occurrences of \emph{Myrcia hatschbachii} (a tree endemic to southern Brazil). The valid occurrences were sourced from Trindade & Marques (2024).
}
\references{
Trindade, W.C.F., Marques, M.C.M., 2023. The Invisible Species: Big Data Unveil Coverage Gaps in the Atlantic Forest Hotspot. Diversity and Distributions 30, e13931. https://doi.org/10.1111/ddi.13931
}
\keyword{datasets}
