% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_for_projection.R
\name{organize_for_projection}
\alias{organize_for_projection}
\title{Organize and structure variables for past and future projections}
\usage{
organize_for_projection(output_dir, models = NULL,
                               variable_names = NULL,
                               categorical_variables = NULL,
                               present_file = NULL,
                               past_files = NULL, past_period = NULL,
                               past_gcm = NULL, future_files = NULL,
                               future_period = NULL, future_pscen = NULL,
                               future_gcm = NULL, fixed_variables = NULL,
                               check_extent = TRUE,
                               resample_to_present = TRUE, mask = NULL,
                               overwrite = FALSE)
}
\arguments{
\item{output_dir}{(character) path to the folder where the organized data
will be saved.}

\item{models}{an object of class fitted_models returned by the
\code{fit_selected()} function. Default is NULL.}

\item{variable_names}{(character) names of the variables used to fit the
model or do the PCA in the \code{prepare_data()} function. Only applicable if
'models' argument is not provided. Default is NULL.}

\item{categorical_variables}{(character) names of the variables that are
categorical. Default is NULL.}

\item{present_file}{(character) \strong{full paths} to the variables from the
present scenario. Default is NULL.}

\item{past_files}{(character) \strong{full paths} to the variables from the past
scenario(s). Default is NULL.}

\item{past_period}{(character) names of the subfolders within 'past_files',
representing specific time periods (e.g., 'LGM' or 'MID'). Only applicable
if 'past_files' is provided. Default is NULL.}

\item{past_gcm}{(character) names of the subfolders within 'past_files',
representing specific General Circulation Models (GCMs). Only applicable if
'past_files' is provided. Default is NULL.}

\item{future_files}{(character) \strong{full paths} to the variables from the
future scenario(s). Default is NULL.}

\item{future_period}{(character) names of the subfolders within
'future_files', representing specific time periods (e.g., '2041-2060' or
'2081-2100'). Only applicable if 'future_files' is provided. Default is NULL.}

\item{future_pscen}{(character) names of the subfolders within 'future_files',
representing specific emission scenarios (e.g., 'ssp126' or 'ssp585'). Only
applicable if 'future_files' is provided. Default is NULL.}

\item{future_gcm}{(character) names of the subfolders within 'future_files',
representing specific General Circulation Models (GCMs). Only applicable if
'future_files' is provided. Default is NULL.}

\item{fixed_variables}{(SpatRaster) optional static variables (i.e., soil
type) used in the model, which will remain unchanged in past or future
scenarios. This variable will be included with each scenario. Default is NULL.}

\item{check_extent}{(logical) whether to ensure that the 'fixed_variables'
have the same spatial extent as the bioclimatic variables. Applicable only if
'fixed_variables' is provided. Default is TRUE.}

\item{resample_to_present}{(logical) whether to resample past or future
variables so they match the extent of the present variables. Only used when
'present_file' is provided. Default is TRUE.}

\item{mask}{(SpatRaster, SpatVector, or SpatExtent) spatial object used to
mask the variables (optional). Default is NULL.}

\item{overwrite}{whether to overwrite existing files in the output directory.
Default is FALSE.}
}
\value{
A message indicating that the variables were successfully organized in the
'output_dir' directory.
}
\description{
This function helpts to organize climate variable files from past and future
scenarios into folders categorized by time period ("Past" or "Future"),
specific period (e.g., "LGM" or "2081–2100"), emission scenario (e.g.,
"ssp585"), and GCMs. This structure simplifies the preparation of climate
data and ensures compatibility with the \code{prepare_projection()} function,
making the variables properly organized for modeling projections.
See \strong{Details} for more information.
}
\details{
The listed input rasters must be stored as \code{.tif} files, with one file per
scenario. Filenames should include identifiable patterns for time period,
GCM, and (for future scenarios) the emission scenario (SSP).

For example:
\itemize{
\item A file representing "Past" conditions for the "LGM" period using the
"MIROC6" GCM should be named: \code{"Past_LGM_MIROC6.tif"}
\item A file representing "Future" conditions for the period "2081–2100" under
the emission scenario "ssp585" and the GCM "ACCESS-CM2" should be named:
\code{"Future_2081-2100_ssp585_ACCESS-CM2.tif"}
}

All scenario files must contain the same variable names (e.g., \code{bio1},
\code{bio2}, etc.) and units as those used for model calibration with present-day
data.
Tip: When listing the files, use \code{list.files(path, full.names = TRUE)} to
obtain the full file paths required by the function.
}
\examples{
# Set the input directory containing the climate variables.
# In this example, we use present and LGM variables from CHELSA
# located in the "inst/extdata" folder of the package.
present_lgm_dir <- system.file("extdata", package = "kuenm2")

# Define an output directory (here, using a temporary folder)
# Replace with your own working directory if needed.
out_dir <- file.path(tempdir(), "Projection_variables")

# List files for present-day conditions
present_list <- list.files(path = present_lgm_dir,
                           pattern = "Current_CHELSA", # Select only CHELSA present-day files
                           full.names = TRUE)

# List files for LGM conditions
lgm_list <- list.files(path = present_lgm_dir,
                       pattern = "LGM", # Select only LGM files
                       full.names = TRUE)

# Organize variables for projection
organize_for_projection(output_dir = out_dir,
                        variable_names = c("bio1", "bio7", "bio12", "bio15"),
                        present_file = present_list,
                        past_files = lgm_list,
                        past_period = "LGM",
                        past_gcm = c("CCSM4", "CNRM-CM5", "FGOALS-g2",
                                     "IPSL-CM5A-LR", "MIROC-ESM", "MPI-ESM-P",
                                     "MRI-CGCM3"),
                        resample_to_present = TRUE,
                        overwrite = TRUE)

}
\seealso{
\link{prepare_projection} \link{organize_future_worldclim}
}
