% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_explore_partition.R
\name{plot_partition_extrapolation}
\alias{plot_partition_extrapolation}
\title{Plots to explore extrapolation risks in data training/testing partitions}
\usage{
plot_partition_extrapolation(
  explore_partition,
  data = NULL,
  space = "G",
  type_of_plot = "mop_distance",
  variables = NULL,
  use_pca = TRUE,
  pcs = c("PC1", "PC2"),
  calibration_area = NULL,
  include_test_background = FALSE,
  distance_pallete = pals::brewer.rdylgn(5),
  custom_pallete = NULL,
  break_type = "pretty",
  in_range_color = "#009E73",
  out_range_color = "#D55E00",
  calibration_area_col = "gray90",
  pr_transparency = 1,
  bg_transparency = 0.4,
  pch_in_range = 19,
  pch_out_range = 18,
  cex_plot = 1.4,
  size_text_legend = 1,
  ncols = NULL
)
}
\description{
Plots to explore extrapolation risks in data training/testing partitions
}
