% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_methods.R
\name{predict.glmnet_mx}
\alias{predict.glmnet_mx}
\title{Predict Maxent-like glmnet models on new data}
\usage{
predict.glmnet_mx(
  object,
  newdata,
  clamp = FALSE,
  type = c("link", "exponential", "cloglog", "logistic")
)
}
\description{
Predict Maxent-like glmnet models on new data
}
