% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{predict}
\alias{predict}
\alias{predict.glmnet_mx}
\alias{predict,kuenm2_glmnet_mx-method}
\title{Predict method for glmnet_mx (maxnet) models}
\usage{
predict.glmnet_mx(object, newdata, clamp = FALSE,
                  type = c("link", "exponential", "cloglog", "logistic",
                  "cumulative"))
}
\arguments{
\item{object}{a glmnet_mx object.}

\item{newdata}{data to predict on.}

\item{clamp}{(logical) whether to clamp predictions. Default = FALSE.}

\item{type}{(character) type of prediction to be performed. Options are:
"link", "exponential", "cloglog", "logistic", and cumulative. Defaults to "link" if not defined.}
}
\value{
A glmnet_mx (maxnet) prediction.
}
\description{
Predict method for glmnet_mx (maxnet) models
}
