% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_selected_glmnetmx.R
\name{predict_selected_glmnetmx}
\alias{predict_selected_glmnetmx}
\title{Predict selected Maxent-like glmnet models}
\usage{
predict_selected_glmnetmx(
  models,
  spat_var,
  do_pca = FALSE,
  write_files = FALSE,
  write_replicates = FALSE,
  out_dir = NULL,
  consensus_per_model = TRUE,
  consensus_general = TRUE,
  consensus = c("median", "range", "mean", "stdev"),
  clamping = FALSE,
  var_to_clamp = NULL,
  type = "cloglog",
  overwrite = FALSE,
  progress_bar = TRUE
)
}
\description{
Predict selected Maxent-like glmnet models
}
