% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_changes.R
\name{prediction_changes}
\alias{prediction_changes}
\title{Compute changes of suitable areas in other scenarios (single scenario / GCM)}
\usage{
prediction_changes(current_predictions, new_predictions,
                   predicted_to = "future", fitted_models = NULL,
                   consensus = "mean", user_threshold = NULL,
                   force_resample = FALSE, gain_color = "#009E73",
                   loss_color = "#D55E00", stable_suitable = "#0072B2",
                   stable_unsuitable = "grey", write_results = FALSE,
                   output_dir = NULL, overwrite = FALSE,
                   write_bin_models = FALSE)
}
\arguments{
\item{current_predictions}{(SpatRaster) A \code{SpatRaster} object returned by
\code{predict_selected()} with suitability predicted under current conditions.}

\item{new_predictions}{(SpatRaster) A \code{SpatRaster} object returned by
\code{predict_selected()} with suitability predicted under new conditions
(past or future). This SpatRaster must have the same resolution and extent
as \code{current_predictions}.}

\item{predicted_to}{(character) a string specifying whether \code{new_predictions}
represent "past" or "future" conditions. Default is "future".}

\item{fitted_models}{an object of class \code{fitted_models} returned by
\code{\link[=fit_selected]{fit_selected()}}}

\item{consensus}{(character) the consensus metric stored in \code{fitted_models}
used to binarize models. Available options are \code{"mean"}, \code{"median"},
\code{"range"}, and \code{"stdev"} (standard deviation). Default is \code{"mean"}.}

\item{user_threshold}{(numeric) an optional threshold for binarizing predictions.
Default is \code{NULL}, meaning the function will apply the thresholds stored in
\code{model_projections}, which were calculated earlier using the omission rate
from \code{\link[=calibration]{calibration()}}.}

\item{force_resample}{(logical) whether to force rasters to have the same
extent and resolution. Default is \code{TRUE}.}

\item{gain_color}{(character) color used to represent gains. Default is
"#009E73" (teal green).}

\item{loss_color}{(character) color used to represent losses. Default is
"#D55E00" (orange-red).}

\item{stable_suitable}{(character) color used for representing areas that
remain suitable across scenarios. Default is "#0072B2" (oxford blue).}

\item{stable_unsuitable}{(character) color used for representing areas that
remain unsuitable across scenarios. Default is "grey".}

\item{write_results}{(logical) whether to save the results to disk. Default
is FALSE.}

\item{output_dir}{(character) directory path where results will be saved.
Only relevant if \code{write_results = TRUE}.}

\item{overwrite}{(logical) whether to overwrite SpatRasters if they already
exist. Only applicable if \code{write_results = TRUE}. Default is FALSE.}

\item{write_bin_models}{(logical) whether to write the binarized models for
each scenario to the disk. Only applicable if \code{write_results = TRUE}. Default
is FALSE.}
}
\value{
A \code{SpatRaster} showing the areas of \strong{gain}, \strong{loss} and \strong{stability}.
}
\description{
Compute changes of suitable areas in other scenarios (single scenario / GCM)
}
\details{
When projecting a niche model to different temporal scenarios (past or
future), species’ areas can be classified into three categories relative to
the current baseline: \strong{gain}, \strong{loss} and \strong{stability}. The
interpretation of these categories depends on the temporal direction of the projection.
\strong{When projecting to future scenarios}:
\itemize{
\item \emph{Gain}: Areas that are currently unsuitable become suitable in the future.
\item \emph{Loss}: Areas that are currently suitable become unsuitable in the future.
\item \emph{Stability}: Areas that retain their current classification in the future,
whether suitable or unsuitable.
}

\strong{When projecting to past scenarios}:
\itemize{
\item \emph{Gain}: Areas that were unsuitable in the past are now suitable in the
present.
\item \emph{Loss}: Areas that were suitable in the past are now unsuitable in the
present.
\item \emph{Stability}: Areas that retain their past classification in the present,
whether suitable or unsuitable.
}
}
\examples{
# Import an example of fitted models (output of fit_selected())
data("fitted_model_maxnet", package = "kuenm2")


# Import current variables for prediction
present_var <- terra::rast(system.file("extdata", "Current_variables.tif",
                                       package = "kuenm2"))


# Import variables for a single future scenario for prediction
future_var <- terra::rast(system.file("extdata",
                                      "wc2.1_10m_bioc_ACCESS-CM2_ssp585_2081-2100.tif",
                                      package = "kuenm2"))

# Rename variables to match the variable names used in the fitted models
names(future_var) <- sub("bio0", "bio", names(future_var))
names(future_var) <- sub("bio", "bio_", names(future_var))

# Append the static soil variable to the future variables
future_var <- c(future_var, present_var$SoilType)

# Predict under present and future conditions
p_present <- predict_selected(models = fitted_model_maxnet,
                              new_variables = present_var)
p_future <- predict_selected(models = fitted_model_maxnet,
                             new_variables = future_var)


# Compute changes between scenarios
p_changes <- prediction_changes(current_predictions = p_present$General_consensus$mean,
                                new_predictions = p_future$General_consensus$mean,
                                fitted_models = fitted_model_maxnet,
                                predicted_to = "future")
# Plot result
terra::plot(p_changes)
}
