% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data for model calibration using basic inputs}
\usage{
prepare_data(
  occ,
  species = NULL,
  x,
  y,
  spat_variables,
  categorical_variables = NULL,
  nbg = 10000,
  kfolds = 4,
  weights = NULL,
  include_xy = TRUE,
  write_files = F,
  file_name = NULL,
  seed = 1,
  verbose = TRUE
)
}
\description{
Prepare data for model calibration using basic inputs
}
