% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_prepare_data.R
\name{prepare_formulas_glm}
\alias{prepare_formulas_glm}
\title{Prepare Formulas for GLM}
\usage{
prepare_formulas_glm(independent, type = "l", categorical_var = NULL)
}
\description{
Prepare Formulas for GLM
}
