% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_prepare_data.R
\name{prepare_formulas_glmnetmx}
\alias{prepare_formulas_glmnetmx}
\title{Prepare Formulas for GLMNET}
\usage{
prepare_formulas_glmnetmx(
  independent,
  type = "lqpht",
  categorical_var = NULL,
  minvar = 1,
  maxvar = NULL
)
}
\description{
Prepare Formulas for GLMNET
}
