% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_occurrences.R
\name{prepare_occurrences}
\alias{prepare_occurrences}
\title{Prepare occurrence data for ecological niche modeling}
\usage{
prepare_occurrences(data, species_column, longitude_column,
                    latitude_column, has_values = NULL,
                    calibration_variables = NULL, remove_na = TRUE,
                    partition_method = c("random", "block"),
                    train_percentage = 75, random_train_iterations = 2,
                    block_rows = 2, block_columns = 2, set_seed = 1)
}
\arguments{
\item{data}{data.frame with clean occurrence records.}

\item{species_column}{(character) name of the column in \code{data}
containing species name.}

\item{longitude_column}{(character) name of the column in \code{data}
containing longitude values.}

\item{latitude_column}{(character) name of the column in \code{data}
containing latitude values.}

\item{has_values}{(logical) whether \code{data} contains columns representing
environmental variables to be used in models. Default = NULL.}

\item{calibration_variables}{\code{\link[terra]{SpatRaster}}) object
containing environmental variables to be used in models. Default =  NULL.
If \code{has_values} = NULL, this argument must be defined.}

\item{remove_na}{(logical) whether to remove rows containing NA values.
Default = TRUE.}

\item{partition_method}{(character) method to partition training and testing
data. Options are "random" and "block".}

\item{train_percentage}{(numeric) percentage of data to be used for training
models. Default = 75.}

\item{random_train_iterations}{(numeric) number of training-testing
iterations. Valid only if \code{partition_method} = "random". Default = 2.}

\item{block_rows}{(numeric) number of rows to be considered for the matrix
that generates blocks. Valid only if \code{partition_method} = "block".
Default = 2.}

\item{block_columns}{(numeric) number of columns to be considered for the
matrix that generates blocks. Valid only if \code{partition_method} = "block".
Default = 2.}

\item{set_seed}{(numeric) seed to be used as a base when running steps
involving random processes. Default = 1.}
}
\value{
An object of class "occ_cal" containing the following elements:
\itemize{
\item summary.- summary of the parameters used to run analysis.
\item occurrences.- data.frame with occurrence data ready for models.
\item block_limits.- data.frame indicating limits used to assign blocks in data.
\item calibration_regime.- list with information to perform model calibration
routines.
}
}
\description{
Set of analyses to prepare occurrence data for ecological niche
modeling.
}
\details{
Number of training-testing iterations when \code{partition_method} = "block"
depends on the number of blocks resulting from partition. In general, 75\% of
the data is used for training with this partition method. All possible
combinations of training and testing blocks are explored to define the final
number of training-testing iterations.
}
