% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_user_data.R
\name{prepare_user_data}
\alias{prepare_user_data}
\title{Prepare data for model calibration with user-prepared calibration data}
\usage{
prepare_user_data(algorithm, user_data, pr_bg = "pr_bg", species = NULL,
                  x = NULL,y = NULL, categorical_variables = NULL,
                  do_pca = FALSE, exclude_from_pca = NULL,
                  variance_explained = 95, min_explained = 5, center = TRUE,
                  scale = FALSE, write_pca = FALSE, pca_directory = NULL,
                  kfolds = 4, weights = NULL, min_number = 2,
                  min_continuous = NULL,
                  features = c("q", "lq", "lp", "qp", "lqp"),
                  reg_mult = c(0.1, 0.5, 1, 2, 3), include_xy = TRUE,
                  write_file = FALSE, file_name = NULL, seed = 1)
}
\arguments{
\item{algorithm}{(character) type algorithm, either "glm" or "glmnet".}

\item{user_data}{(data frame) A data.frame with columns indicating presence (1)
and background (0) data, as well as the corresponding predictor values. For
an example, see \code{data("user_data", package = "kuenm2")}.}

\item{pr_bg}{(character) a string specifying the name of the column in
user_data that contains the presence/background.}

\item{species}{(character) string specifying the species name (optional).
Default is NULL.}

\item{x}{(character) a string specifying the name of the column in \code{occ} that
contains the longitude values.}

\item{y}{(character) a string specifying the name of the column in \code{occ} that
contains the latitude values.}

\item{categorical_variables}{(character) names of the variables that are
categorical. Default is NULL.}

\item{do_pca}{(logical) whether to perform a principal component analysis
(PCA) with the set of variables. Default is FALSE.}

\item{exclude_from_pca}{(character) variable names within raster_variables
that should not be included in the PCA transformation. Instead, these
variables will be added directly to the final set of output variables
without being modified. The default is NULL, meaning all variables
will be used unless specified otherwise.}

\item{variance_explained}{(numeric) the cumulative percentage of total
variance that must be explained by the selected principal components.
Default is 95.}

\item{min_explained}{(numeric) the minimum percentage of total variance that
a principal component must explain to be retained. Default is 5.}

\item{center}{(logical) whether the variables should be zero-centered. Default
is TRUE.}

\item{scale}{(logical) whether the variables should be scaled to have unit
variance before the analysis takes place. Default is FALSE.}

\item{write_pca}{(logical) whether to save the PCA-derived raster layers
(principal components) to disk. Default is FALSE.}

\item{pca_directory}{(character) the path or name of the folder where the PC
raster layers will be saved. This is only applicable if \code{write_pca = TRUE}.
Default is NULL.}

\item{kfolds}{(numeric) the number of groups (folds) the occurrence
data will be split into for cross-validation. Default is 4.}

\item{weights}{(numeric) a numeric vector specifying weights for the
occurrence records. Default is NULL.}

\item{min_number}{(numeric) the minimum number of variables to be included in
the model formulas to be generated.}

\item{min_continuous}{(numeric) the minimum number of continuous variables
required in a combination. Default is NULL.}

\item{features}{(character) a vector of feature classes. Default is c("q",
"lq", "lp", "qp", "lqp").}

\item{reg_mult}{(numeric) a vector of regularization parameters for glmnet.
Default is c(0.1, 1, 2, 3, 5).}

\item{include_xy}{(logical) whether to include the coordinates (longitude and
latitude) in the results from preparing data. Default is TRUE.}

\item{write_file}{(logical) whether to write the resulting prepared_data list
in a local directory. Default is FALSE.}

\item{file_name}{(character) the path or name of the folder where the
resulting list will be saved. This is only applicable if \code{write_file = TRUE}. Default is NULL.}

\item{seed}{(numeric) integer value to specify an initial seed to split the
data. Default is 1.}
}
\value{
An object of class \code{prepared_data} containing all elements to run a model
calibration routine. The elements include: species, calibration data,
a grid of model parameters, indices of k-folds for cross validation,
xy coordinates, names of continuous and categorical variables, weights,
results from PCA, and modeling algorithm.
}
\description{
This function prepares data for model calibration using user-prepared
calibration data. It includes optional PCA, k-fold partitioning, and the
creation of a grid parameter combinations, including distinct regularization
multiplier values, various feature classes, and different sets of
environmental variables.
}
\examples{
# Import user-prepared calibration data
data("user_data", package = "kuenm2")
# Prepare data for glmnet model
glmnet_swd_user <- prepare_user_data(algorithm = "glmnet", user_data = user_data,
                                     species = "Myrcia hatschbachii",
                                     categorical_variables = "SoilType",
                                     features = c("l", "q", "p", "lq", "lqp"),
                                     reg_mult = c(0.1, 1, 2, 3, 5))
print(glmnet_swd_user)

# Prepare data for glm model
glm_swd_user <- prepare_user_data(algorithm = "glm", user_data = user_data,
                                  species = "Myrcia hatschbachii",
                                  categorical_variables = "SoilType",
                                  features = c("l", "q", "p", "lq", "lqp"),
                                  reg_mult = c(0.1, 1, 2, 3, 5))
print(glm_swd_user)
}
