% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.model_projections}
\alias{print.model_projections}
\title{Print Method for model_projections Class}
\usage{
\method{print}{model_projections}(x, ...)
}
\description{
Print Method for model_projections Class
}
