% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{print}
\alias{print}
\alias{print.prepared_data}
\alias{print,kuenm2_prepared_data-method}
\alias{print,kuenm2_calibration_results-method}
\alias{print,kuenm2_fitted_models-method}
\alias{print,kuenm2_projection_data-method}
\alias{print,kuenm2_model_projections-method}
\alias{print.calibration_results}
\alias{print.fitted_models}
\alias{print.projection_data}
\alias{print.model_projections}
\title{Print method for kuenm2 objects}
\usage{
\method{print}{prepared_data}(x, ...)

\method{print}{calibration_results}(x, ...)

\method{print}{fitted_models}(x, ...)

\method{print}{projection_data}(x, ...)

\method{print}{model_projections}(x, ...)
}
\arguments{
\item{x}{an object of any of these classes: \code{prepared_data},
\code{calibration_results}, \code{fitted_models}, \code{projection_data}, or
\code{model_projections}.}

\item{...}{additional arguments affecting the summary produced. Ignored in
these functions.}
}
\value{
A printed version of the object that summarizes the main elements contained.
}
\description{
Print method for kuenm2 objects
}
