% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_changes.R
\name{proj_changes}
\alias{proj_changes}
\title{Title}
\usage{
proj_changes(
  projection_paths,
  referece_id = 1,
  consensus = "median",
  include_id = NULL,
  user_thr = NULL,
  by_gcm = TRUE,
  by_change = TRUE,
  general_summary = TRUE,
  force_resample = TRUE,
  write_results = TRUE,
  output_dir = NULL,
  overwrite = TRUE,
  write_bin_models = FALSE,
  return_rasters = FALSE
)
}
\arguments{
\item{return_rasters}{}
}
\description{
Title
}
