% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_selected_glmnetx.R
\name{project_selected_glmnetx}
\alias{project_selected_glmnetx}
\title{Projection of Maxent-like glmnet models to single or multiple scenarios}
\usage{
project_selected_glmnetx(
  models,
  projection_file,
  out_dir = "Projection_results",
  write_path = TRUE,
  consensus_per_model = TRUE,
  consensus_general = TRUE,
  consensus = c("median", "range", "mean", "stdev"),
  write_replicates = FALSE,
  clamping = FALSE,
  var_to_clamp = NULL,
  type = "cloglog",
  overwrite = FALSE,
  progress_bar = TRUE,
  verbose = TRUE
)
}
\description{
Projection of Maxent-like glmnet models to single or multiple scenarios
}
