% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_curves.R
\name{response_curve}
\alias{response_curve}
\title{Variable response curves for fitted models}
\usage{
response_curve(models, variable, modelID = NULL, n = 100,
               show_variability = FALSE, show_lines = FALSE, data = NULL,
               new_data = NULL, averages_from = "pr_bg", extrapolate = TRUE,
               extrapolation_factor = 0.1, add_points = FALSE, p_col = NULL,
               l_limit = NULL, u_limit = NULL,
               xlab = NULL, ylab = "Suitability",
               col = "darkblue", ...)
}
\arguments{
\item{models}{an object of class \code{fitted_models} returned by the
\code{\link[=fit_selected]{fit_selected()}} function.}

\item{variable}{(character) name of the variable to be plotted.}

\item{modelID}{(character) vector of ModelID(s) to be considered in the
models object. By default all models are included. Default = NULL.}

\item{n}{(numeric) an integer guiding the number of breaks. Default = 100}

\item{show_variability}{(logical) if \code{modelID} is defined, shows variability
in response curves considering replicates. If \code{modelID} is not defined, the
default, FALSE, always shows variability from multiple models if present in
\code{models}.}

\item{show_lines}{(logical) whether to show variability by plotting lines for
all models or replicates. The default = FALSE, uses a GAM to characterize a
median trend and variation among modes or replicates. Ignored if
\code{show_variability} = FALSE.}

\item{data}{data.frame or matrix of data used in the model calibration step.
Default = NULL.}

\item{new_data}{a \code{SpatRaster}, data.frame, or  matrix of variables
representing the range of variable values in an area of interest.
Default = NULL. It must be defined in case the model entered does not
explicitly include a data component.}

\item{averages_from}{(character) specifies how the averages or modes of the
variables are calculated. Available options are "pr" (to calculate averages
from the presence localities) or "pr_bg" (to use the combined set of presence
and background localities). Default is "pr_bg". See details.}

\item{extrapolate}{(logical) whether to allow extrapolation to study the
behavior of the response outside the calibration limits. Ignored if
\code{new_data} is defined. Default = TRUE.}

\item{extrapolation_factor}{(numeric) a multiplier used to calculate the
extrapolation range. Larger values allow broader extrapolation beyond the
observed data range. Default is 0.1.}

\item{add_points}{(logical) if \code{TRUE}, adds the original observed
points (0/1) to the plot. This also sets \code{ylim = c(0, 1)}, unless these
limits are defined as part of \code{...}. Default = \code{FALSE}.}

\item{p_col}{(character) color for the observed points when
\code{add_points = TRUE}. Any valid R color name or hexadecimal code.
Default = "black".}

\item{l_limit}{(numeric) specifies the lower limit for the variable. Default
is \code{NULL}, meaning the lower limit will be calculated based on the
data's minimum value and the \code{extrapolation_factor}
(if \code{extrapolation = TRUE}).}

\item{u_limit}{(numeric) specifies the upper limit for the variable. Default
is \code{NULL}, meaning the upper limit will be calculated based on the
data's minimum value and the \code{extrapolation_factor}
(if \code{extrapolation = TRUE}).}

\item{xlab}{(character) a label for the x axis. The default, NULL, uses the
name defined in \code{variable}.}

\item{ylab}{(character) a label for the y axis. Default = "Suitability".}

\item{col}{(character) color for lines. Default = "darkblue".}

\item{...}{additional arguments passed to \code{\link[graphics]{plot}}.}
}
\value{
A plot with the response curve for a \code{variable}.
}
\description{
A view of variable responses in fitted models. Responses based on single or
multiple models can be plotted.
}
\details{
The response curves are generated with all other variables set to their mean
values (or mode for categorical variables), calculated either from the
presence localities (if averages_from = "pr") or from the combined set of
presence and background localities (if averages_from = "pr_bg").

For categorical variables, a bar plot is generated with error bars showing
variability across models (if multiple models are included).
}
\examples{
# Example with maxnet
# Import example of fitted_models (output of fit_selected())
data(fitted_model_maxnet, package = "kuenm2")

#Response curves
response_curve(models = fitted_model_maxnet, variable = "bio_1",
               show_variability = TRUE)
response_curve(models = fitted_model_maxnet, variable = "bio_1",
               show_variability = TRUE, add_points = TRUE)
response_curve(models = fitted_model_maxnet, variable = "bio_1",
               show_variability = TRUE, show_lines = TRUE)
response_curve(models = fitted_model_maxnet, variable = "bio_1",
               modelID = "Model_192", show_variability = TRUE)
response_curve(models = fitted_model_maxnet, variable = "bio_1",
               modelID = "Model_192", show_variability = TRUE,
               show_lines = TRUE)

# Example with GLM
# Import example of fitted_models (output of fit_selected())
data(fitted_model_glm, package = "kuenm2")

#Response curves
response_curve(models = fitted_model_glm,
               variable = "bio_1", show_variability = TRUE)
response_curve(models = fitted_model_glm, variable = "bio_1",
               modelID = "Model_85", show_variability = TRUE)
}
\seealso{
\code{\link[=bivariate_response]{bivariate_response()}}
}
