% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_curves.R
\name{response_curve_glmnetmx}
\alias{response_curve_glmnetmx}
\title{Response curve}
\usage{
response_curve_glmnetmx(
  fitted,
  variable,
  modelID = NULL,
  n = 100,
  by_replicates = FALSE,
  data = NULL,
  new_data = NULL,
  extrapolate = TRUE,
  xlab = NULL,
  ylab = "Suitability",
  col = "darkblue",
  ...
)
}
\description{
Response curve
}
