% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sel_best_models.R
\name{sel_best_models}
\alias{sel_best_models}
\title{Select best among candidate models}
\usage{
sel_best_models(
  cand_models,
  test_concave = TRUE,
  omrat_threshold = 5,
  allow_tolerance = T,
  tolerance = 0.01,
  AIC = "nk",
  significance = 0.05,
  verbose = TRUE,
  delta_aic = 2,
  save_file = T,
  file_name = NULL
)
}
\description{
Select best among candidate models
}
