% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_models.R
\name{select_models}
\alias{select_models}
\title{Select models that perform the best among candidates}
\usage{
select_models(calibration_results = NULL, candidate_models = NULL, data = NULL,
              algorithm = NULL, compute_proc = FALSE,
              addsamplestobackground = TRUE, weights = NULL,
              test_concave = FALSE, omrat_threshold = 10,
              allow_tolerance = TRUE, tolerance = 0.01, AIC_option = "ws",
              significance = 0.05, delta_aic = 2, parallel = FALSE,
              ncores = NULL, progress_bar = FALSE,verbose = TRUE)
}
\arguments{
\item{calibration_results}{an object of class \code{calibration_results} returned
by the \code{\link[=calibration]{calibration()}} function. Default is NULL.}

\item{candidate_models}{(data.frame) a summary of the evaluation metrics for each
candidate model. Required only if \code{calibration_results} is NULL. In the
output of the \code{\link[=calibration]{calibration()}}, this data.frame is located in
\verb{$calibration_results$Summary}. Default is NULL.}

\item{data}{an object of class \code{prepared_data} returned by the
\code{\link[=prepare_data]{prepare_data()}} function. Required only if \code{calibration_results}
is NULL and \code{compute_proc} is TRUE.}

\item{algorithm}{(character) model algorithm, either "glm" or "maxnet". The
default, NULL, uses the one defined as part of \code{calibration_results}, or
\code{data}. If those arguments are not used, \code{algorithm} must be defined.}

\item{compute_proc}{(logical) whether to compute partial ROC tests for the
selected models. This is required when partial ROC is not calculated for all
candidate models during calibration. Default is FALSE.}

\item{addsamplestobackground}{(logical) whether to add to the background any
presence sample that is not already there. Required only if \code{compute_proc} is
TRUE and \code{calibration_results} is NULL.Default is TRUE.}

\item{weights}{(numeric) a numeric vector specifying weights for the
occurrence records. Required only if \code{compute_proc} is TRUE and
\code{calibration_results} is NULL. Default is NULL.}

\item{test_concave}{(logical) whether to remove candidate models presenting
concave curves. Default is FALSE.}

\item{omrat_threshold}{(numeric) the maximum omission rate a candidate model
can have to be considered a best model. Default is 10. This value must match
one of the values specified in \code{omrat} in \code{\link[=calibration]{calibration()}}.}

\item{allow_tolerance}{(logical) whether to allow selection of models with
minimum values of omission rates even if their omission rate surpasses the
\code{omrat_threshold}. This is only applicable if all candidate models have
omission rates higher than the \code{omrat_threshold}. Default is TRUE.}

\item{tolerance}{(numeric) The value added to the minimum omission rate if it
exceeds the \code{omrat_threshold}. If \code{allow_tolerance = TRUE}, selected models
will have an omission rate equal to or less than the minimum rate plus this
tolerance. Default is 0.01.}

\item{AIC_option}{(character) the type of AIC to be calculated: "ws" for AIC
proposed by Warren and Seifert (2011), or "nk" for AIC proposed by Ninomiya
and Kawano (2016). This is only applicable if algorithm = "maxnet".
Default is "ws". See References for details.}

\item{significance}{(numeric) the significance level to select models
based on the partial ROC (pROC). Default is 0.05. See Details.}

\item{delta_aic}{(numeric) the value of delta AIC used as a threshold to
select models. Default is 2.}

\item{parallel}{(logical) whether to calculate the PROC of the candidate
models in parallel. Default is FALSE.}

\item{ncores}{(numeric) number of cores to use for parallel processing.
Default is NULL and uses available cores - 1. This is only applicable if
\code{parallel = TRUE}.}

\item{progress_bar}{(logical) whether to display a progress bar during
processing. Default is TRUE.}

\item{verbose}{(logical) whether to display messages during processing.
Default is TRUE.}
}
\value{
If calibration_results is provided, it returns a new calibration_results with
the new selected models and summary. If calibration_results is NULL, it
returns a list containing the following elements:
\itemize{
\item selected_models: data frame with the ID and the summary of evaluation
metrics for the selected models.
\item summary: A list containing the delta AIC values for model selection, and
the ID values of models that failed to fit, had concave curves,
non-significant pROC values, omission rates above the threshold, delta AIC
values above the threshold, and the selected models.
}
}
\description{
This function selects the best models according to user-defined criteria,
evaluating statistical significance (partial ROC), predictive ability
(omission rates), and model complexity (AIC).
}
\details{
Partial ROC is calculated following Peterson et al.
(2008).
}
\examples{
# Import example of calibration results (output of calibration function)
## GLM
data(calib_results_glm, package = "kuenm2")

#Select new best models based on another value of omission rate
new_best_model <- select_models(candidate_models = calib_results_glm$calibration_results$Summary,
                                algorithm = "glm",
                                omrat_threshold = 5)  # Omission error of 5

# Compare with best models selected previously
calib_results_glm$summary$Selected  # Model 1 selected
new_best_model$summary$Selected  # Models 1 and 5 selected
}
