% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_mop.R
\name{single_mop}
\alias{single_mop}
\title{Analysis of extrapolation risks using the MOP metric (for single scenario)}
\usage{
single_mop(data, new_variables, subset_variables = FALSE,
           mask = NULL, type = "basic", na_in_range = TRUE,
           calculate_distance = FALSE, where_distance = "in_range",
           distance = "euclidean", scale = FALSE, center = FALSE,
           fix_NA = TRUE, percentage = 1, comp_each = 2000, tol = NULL,
           rescale_distance = FALSE, parallel = FALSE, ncores = NULL,
           progress_bar = TRUE, write_files = FALSE, out_dir = NULL,
           overwrite = FALSE)
}
\arguments{
\item{data}{an object of class \code{fitted_models} returned by the
\code{\link[=fit_selected]{fit_selected()}} function, an object of class \code{prepared_data} returned by
the \code{\link[=prepare_data]{prepare_data()}} function, or an object of class \code{calibration_results}
returned by the \code{\link[=calibration]{calibration()}} function.}

\item{new_variables}{a SpatRaster or data.frame of predictor variables.
The names of these variables must match those used to prepare the date or
calibrate the models provided in \code{data}.}

\item{subset_variables}{(logical) whether to include in the analysis only the
variables present in the selected models. Only applicable if \code{data} is a
\code{fitted_models} object. Default is FALSE}

\item{mask}{(SpatRaster, SpatVector, or SpatExtent) spatial object used to
mask the variables (optional). Default is NULL.}

\item{type}{(character) type of MOP analysis to be performed. Options
available are "basic", "simple" and "detailed". See Details for further
information.}

\item{na_in_range}{(logical) whether to assign \code{NA} to regions within the
projected area (G) where environmental conditions fall within the range of
the calibration data (M). If \code{TRUE} (default), these regions are assigned
\code{NA}. If \code{FALSE}, they are assigned \code{0} in the simple and basic MOP outputs,
and \code{"within ranges"} in the detailed MOP output.}

\item{calculate_distance}{(logical) whether to calculate distances
(dissimilarities) between m and g. The default, FALSE, runs rapidly and does
not assess dissimilarity levels.}

\item{where_distance}{(character) where to calculate distances, considering
how conditions in g are positioned in comparison to the range of conditions
in m. Options available are "in_range", "out_range" and "all". Default is
"in_range".}

\item{distance}{(character) which distances are calculated, euclidean or
mahalanobis. Only applicable if calculate_distance = TRUE.}

\item{scale}{(logical or numeric) whether to scale as in
\code{\link[base]{scale}}. Default is FALSE.}

\item{center}{(logical or numeric) whether to center as in
\code{\link[base]{scale}}. Default is FALSE.}

\item{fix_NA}{(logical) whether to fix layers so cells with NA values are
the same in all layers. Setting to FALSE may save time if the rasters are
big and have no NA matching problems. Default is TRUE.}

\item{percentage}{(numeric) percentage of \code{m} closest conditions used
to derive mean environmental distances to each combination of conditions in
\code{g}.}

\item{comp_each}{(numeric) number of combinations in \code{g} to be used for
distance calculations at a time. Increasing this number requires more RAM}

\item{tol}{(numeric) tolerance to detect linear dependencies when calculating
Mahalanobis distances. The default, NULL, uses \code{.Machine$double.eps}.}

\item{rescale_distance}{(logical) whether to re-scale distances 0-1.
Re-scaling prevents comparisons of dissimilarity values obtained from runs
with different values of \code{percentage}.}

\item{parallel}{(logical) whether to compute MOP in parallel. Default is
FALSE.}

\item{ncores}{(numeric) number of cores to use for parallel processing.
Default is NULL and uses available cores - 1. This is only applicable if
\code{parallel = TRUE}.}

\item{progress_bar}{(logical) whether to display a progress bar during
distance calculation. Only applicable if calculate_distance is TRUE.
Default is TRUE.}

\item{write_files}{(logical) whether to save the MOP results (SpatRasters and
data.frames) to disk. Default is FALSE.}

\item{out_dir}{(character) directory path where results will be saved.
Only relevant if \code{write_files = TRUE}.}

\item{overwrite}{(logical) whether to overwrite SpatRasters if they already
exist. Only applicable if \code{write_files = TRUE}. Default is FALSE.}
}
\value{
An object of class \code{mop_results} containing:
\itemize{
\item \strong{summary} - a data.frame with details of the data used in the analysis:
\itemize{
\item \emph{variables} - names of variables considered.
\item \emph{type} - type of MOP analysis performed.
\item \emph{scale} - value according to the argument \code{scale}.
\item \emph{center} - value according to the argument \code{center}.
\item \emph{calculate_distance} - value according to the argument
\code{calculate_distance}.
\item \emph{distance} - option regarding distance used.
\item \emph{percentage} - percentage of \code{m} used as reference for
distance calculation.
\item \emph{rescale_distance} - value according to the argument
\code{rescale_distance}.
\item \emph{fix_NA} - value according to the argument \code{fix_NA}.
\item \emph{N_m} - total number of elements (cells with values or valid
rows) in \code{m}.
\item \emph{N_g} - total number of elements (cells with values or valid
rows) in \code{g}.
\item \emph{m_ranges} - the range (minimum and maximum values) of the variable in
reference conditions (\code{m})
}
\item \strong{mop_distances} - if \code{calculate_distance} = TRUE, a SpatRaster
with distance values for the set of interest (\code{g}). Higher values
represent greater dissimilarity compared to the set of reference (\code{m}).
\item \strong{mop_basic} - a SpatRaster for the set of interest representing
conditions in which at least one of the variables is non-analogous to the set
of reference. Values should be: 1 for non-analogous conditions, and NA for
conditions inside the ranges of the reference set.
\item \strong{mop_simple} - a SpatRaster, for the set of interest, representing how
many variables in the set of interest are non-analogous to those in the
reference set. NA is used for conditions inside the ranges of the reference
set.
\item \strong{mop_detailed} - a list containing:
\itemize{
\item \emph{interpretation_combined} - a data.frame to help identify combinations
of variables in \emph{towards_low_combined} and \emph{towards_high_combined} that
are non-analogous to \code{m}.
\item \emph{towards_low_end} - a SpatRaster for all variables representing where
non-analogous conditions were found towards low values of each variable.
\item \emph{towards_high_end} - a SpatRaster for all variables representing where
non-analogous conditions were found towards high values of each variable.
\item \emph{towards_low_combined} - a SpatRaster with values representing the
identity of the variables found to have non-analogous conditions towards
low values.
\item \emph{towards_high_combined} - a SpatRaster with values representing the
identity of the variables found to have non-analogous conditions towards
high values.
}
}
}
\description{
Calculates the mobility-oriented parity metric and other sub-products to
represent dissimilarities and non-analogous conditions when comparing a set
of reference conditions (M) against another set of scenario conditions (G).
}
\details{
\code{type} options return results that differ in the detail of how non-analogous
conditions are identified.
\itemize{
\item \strong{basic} - makes calculation as proposed by Owens et al. (2013)
\url{doi:10.1016/j.ecolmodel.2013.04.011}.
\item \strong{simple} - calculates how many variables in the set of interest are
non-analogous to those in the reference set.
\item \strong{detailed} - calculates five additional extrapolation metrics. See
\code{mop_detailed} under \code{Value} below for full details.
}

\code{where_distance} options determine what values should be used to calculate
dissimilarity
\itemize{
\item \strong{in_range} - only conditions inside \code{m} ranges
\item \strong{out_range} - only conditions outside \code{m} ranges
\item \strong{all} - all conditions
}

When the variables used to represent conditions have different units,
scaling and centering are recommended. This step is only valid when Euclidean
distances are used.
}
\examples{
# Import an example of fitted models (output of fit_selected())
data("fitted_model_maxnet", package = "kuenm2")

# Import variables under a new set of conditions
# Here, future climate data
future_scenario <- terra::rast(system.file("extdata",
                                          "wc2.1_10m_bioc_ACCESS-CM2_ssp585_2081-2100.tif",
                                          package = "kuenm2"))



# Rename variables to match the variable names in the fitted models
names(future_scenario) <- sub("bio0", "bio", names(future_scenario))
names(future_scenario) <- sub("bio", "bio_", names(future_scenario))

# Run MOP
sm <- single_mop(data = fitted_model_maxnet, new_variables = future_scenario,
                 type = "detailed")
}
\seealso{
\code{\link[=projection_mop]{projection_mop()}}
}
