% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp_swd_glm}
\alias{sp_swd_glm}
\title{Prepared Data for glm models}
\format{
A \code{prepared_data} object with the following elements:
\describe{
\item{species}{Species names}
\item{calibration_data}{A \code{data.frame} containing the variables extracted for presence and background points}
\item{formula_grid}{A \code{data.frame} with the ID, formulas, and regularization multipliers of each candidate model}
\item{part_data}{A \code{list} with the partition data, where each element corresponds to a replicate and contains the \strong{indices of the test points} for that replicate}
\item{partition_method}{A \code{character} indicating the partition method}
\item{n_replicates}{A \code{numeric} value indicating the number of replicates or k-folds}
\item{train_proportion}{A \code{numeric} value indicating the proportion of occurrences used as train points when the partition method is 'subsample' or 'boostrap'}
\item{data_xy}{A \code{data.frame} with the coordinates of the occurrence and bakground points}
\item{continuous_variables}{A \code{character} indicating the names of the continuous variables}
\item{categorical_variables}{A \code{character} indicating the names of the categorical variables}
\item{weights}{A \code{numeric} value specifying weights for the occurrence records. It's NULL, meaning it was not set weights.}
\item{pca}{A \code{prcomp} object storing PCA information. Is NULL, meaning PCA was not performed}
\item{algorithm}{A \code{character} indicanting the algorithm (glm)}
}
}
\usage{
data("sp_swd_glm")
}
\description{
A \code{prepared_data} object resulted from \code{prepare_data()} to calibrate models using 'glm' algorithm.
}
\keyword{datasets}
