% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swd_spatial_block}
\alias{swd_spatial_block}
\title{Prepared data with spatial blocks created with ENMeval}
\format{
An object of class \code{prepared_data} of length 13.
}
\usage{
data("sp_swd")
}
\description{
A \code{prepared_data} object resulted from \code{prepare_data()} to calibrate
models using 'glmnet' algorithm. In this object, the original partitioning
was replaced with spatial blocks generated using the \code{get.block()}
method from the ENMeval R package.
}
\keyword{datasets}
