% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{user_data}
\alias{user_data}
\title{User Custom Calibration Data}
\format{
A \code{data.frame} with the following columns:
\describe{
\item{pr_bg}{Column indicating presences (1) and background (0).}
\item{bio_1}{The extracted values for the variable bio_1 at presence and background points.}
\item{bio_7}{The extracted values for the variable bio_12 at presence and background points.}
\item{bio_12}{The extracted values for the variable bio_12 at presence and background points.}
\item{bio_15}{The extracted values for the variable bio_15 at presence and background points.}
\item{bio_15}{The extracted values for the variable soilType at presence and background points.}
}
}
\usage{
data("user_data")
}
\description{
A \code{data.frame} containing presence and background records along with environmental variables used to demonstrate data preparation with user-supplied data.
}
\keyword{datasets}
