% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{var}
\alias{var}
\title{SpatRaster Representing present-day Conditions (WorldClim)}
\format{
A \code{SpatRaster} object.
}
\value{
No return value. Used with function \code{\link[terra]{rast}} to
bring raster variables to analysis.
}
\description{
Raster layer containing bioclimatic variables representing present-day
climatic conditions. The variables were obtained at a 10 arc-minute
resolution and masked using the \code{m} region provided in the package. Data
sourced from WorldClim:
\url{https://worldclim.org/data/worldclim21.html}
}
\examples{
var <- terra::rast(system.file("extdata",
                               "Current_variables.tif",
                                package = "kuenm2"))
terra::plot(var)
}
