% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_helpers.R
\name{variable_combinations}
\alias{variable_combinations}
\title{Helper to create all possible sets of variable combinations}
\usage{
variable_combinations(
  variable_names,
  minimum_number = 2,
  maximum_number = NULL
)
}
\arguments{
\item{variable_names}{(character) vector of variable names.}

\item{minimum_number}{(numeric) minimum number of variables per set.}

\item{maximum_number}{(numeric) maximum number of variables per set.}
}
\value{
A list of vectors containing variable names per set.
}
\description{
Helper to create all possible sets of variable combinations
}
