% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_importance.R
\name{variable_importance}
\alias{variable_importance}
\title{Variable importance}
\usage{
variable_importance(models, modelID = NULL, parallel = FALSE, ncores = NULL,
                    progress_bar = TRUE, verbose = TRUE)
}
\arguments{
\item{models}{an object of class \code{fitted_models} returned by the
\code{\link{fit_selected}}() function.}

\item{modelID}{(character). Default = NULL.}

\item{parallel}{(logical) whether to calculate importance in parallel.
Default is FALSE.}

\item{ncores}{(numeric) number of cores to use for parallel processing.
Default is NULL and uses available cores - 1. This is only applicable if
\code{parallel = TRUE}.}

\item{progress_bar}{(logical) whether to display a progress bar during processing.
Default is TRUE.}

\item{verbose}{(logical) whether to display detailed messages during processing.
Default is TRUE.}
}
\value{
A data.frame containing the relative contribution of each variable. An
identification for distinct models is added if \code{fitted} contains multiple
models.
}
\description{
Variable importance
}
\examples{
# Example with maxnet
# Import example of fitted_models (output of fit_selected())
data(fitted_model_maxnet, package = "kuenm2")

# Variable importance
imp_maxnet <- variable_importance(models = fitted_model_maxnet)

# Plot
plot_importance(imp_maxnet)

# Example with glm
# Import example of fitted_models (output of fit_selected())
data(fitted_model_glm, package = "kuenm2")

# Variable importance
imp_glm <- variable_importance(models = fitted_model_glm)

# Plot
plot_importance(imp_glm)
}
\seealso{
\code{\link[=plot_importance]{plot_importance()}}
}
