% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{world}
\alias{world}
\title{World country polygons from Natural Earth}
\format{
A \code{Spatvector} object.
}
\value{
No return value. Used with function \code{\link[terra]{vect}} to
bring vector variables to analysis.
}
\description{
A spatial vector of the world countries. This is a simplified version of the \code{countries110} from rnaturalearth R package.
}
\examples{
m <- terra::vect(system.file("extdata",
                             "world.gpkg",
                              package = "kuenm2"))
terra::plot(m)
}
