% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_cleaning.R
\name{advanced_cleaning}
\alias{advanced_cleaning}
\alias{remove_cell_duplicates}
\alias{move_2closest_cell}
\title{Advanced occurrence data cleaning}
\usage{
advanced_cleaning(data, x, y, raster_layer, cell_duplicates = TRUE,
                  move_points_inside = FALSE, move_limit_distance = NULL,
                  verbose = TRUE)

remove_cell_duplicates(data, x, y,
                       raster_layer)

move_2closest_cell(data, x, y, raster_layer,
                   move_limit_distance, verbose = TRUE)
}
\arguments{
\item{data}{data.frame with occurrence records. Rows with NA values will be
omitted.}

\item{x}{(character) name of the column in \code{data}
containing longitude values.}

\item{y}{(character) name of the column in \code{data}
containing latitude values.}

\item{raster_layer}{a raster layer (object of class
\code{\link[terra]{SpatRaster}}).}

\item{cell_duplicates}{(logical) whether to remove duplicate coordinates
considering raster cells. Default = TRUE.}

\item{move_points_inside}{(logical) whether to move records outside of raster
cells with valid values to the closest cell with values. Default = FALSE.}

\item{move_limit_distance}{maximum distance to move records outside cells
with valid values. Default = NULL. Must be defined if \code{move_points_inside}
= TRUE.}

\item{verbose}{(logical) whether to print messages of progress. Default =
TRUE.}
}
\value{
A data.frame with occurrence records resulting from advanced cleaning
procedures. Other columns will be added to describe changes made in the
original data.
}
\description{
Advanced processes of data cleaning involving duplicate removal
and movement of records.
}
\details{
Data used in this functions should have gone through initial processes of
cleaning and filtering.
}
\examples{
# Import occurrences
data(occ_data_noclean, package = "kuenm2")

# Import raster layers
var <- rast(system.file("extdata", "Current_variables.tif", package = "kuenm2"))

# Keep only one layer
var <- var$bio_1

# all basic cleaning steps
clean_init <- initial_cleaning(data = occ_data_noclean, species = "species",
                               x = "x", y = "y", remove_na = TRUE,
                               remove_empty = TRUE, remove_duplicates = TRUE,
                               by_decimal_precision = TRUE,
                               decimal_precision = 2)

# Advanced cleaning steps
# exclude duplicates based on raster cell (pixel)
celldup <- remove_cell_duplicates(data = clean_init, x = "x", y = "y",
                                  raster_layer = var)

# move records to valid pixels
moved <- move_2closest_cell(data = celldup, x = "x", y = "y",
                            raster_layer = var, move_limit_distance = 10)

# the steps at a time
clean_data <- advanced_cleaning(data = clean_init, x = "x", y = "y",
                                raster_layer = var, cell_duplicates = TRUE,
                                move_points_inside = TRUE,
                                move_limit_distance = 10)
}
\seealso{
\code{\link[=initial_cleaning]{initial_cleaning()}}
}
