% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_prepare_data.R
\name{calibration_grid_glmnetmx}
\alias{calibration_grid_glmnetmx}
\title{Calibration Grid Generation using GLMNET}
\usage{
calibration_grid_glmnetmx(
  occ_bg,
  categorical_var = NULL,
  features = c("l", "q", "lq", "lqp", "p"),
  min_number = 2,
  min_continuous = NULL,
  regm = c(0.1, 1, 2, 3, 5)
)
}
\description{
Calibration Grid Generation using GLMNET
}
