% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_calibration_geo.R
\name{explore_calibration_geo}
\alias{explore_calibration_geo}
\title{Explore the spatial distribution of occurrence and background points}
\usage{
explore_calibration_geo(data, raster_variables, plot = TRUE, ...)
}
\arguments{
\item{data}{an object of class \code{prepared_data} returned by the
\code{\link[=prepare_data]{prepare_data()}} function.}

\item{raster_variables}{(SpatRaster) predictor variables used for model
calibration.}

\item{plot}{(logical) wheter to plot the SpatRaster. Default is TRUE.}

\item{...}{additional arguments passed to \code{terra::plot()}.}
}
\value{
A categorical \code{SpatRaster} with four factor values representing:
\describe{
\item{1 - Background cells}{}
\item{2 - Presence cells}{}
\item{3 - Cells with both presence and background}{}
\item{4 - Non-used cells}{}
}
}
\description{
Explore the spatial distribution of occurrence and background points
}
\examples{
# Import raster layers
var <- terra::rast(system.file("extdata", "Current_variables.tif",
                               package = "kuenm2"))

# Import occurrences
data(sp_swd, package = "kuenm2")

# Explore calibration data
pbg <- explore_calibration_geo(data = sp_swd, raster_variables = var[[1]])
}
