% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_calibration_geo.R
\name{explore_calibration_geo}
\alias{explore_calibration_geo}
\title{Explore the spatial distribution of occurrence and background points}
\usage{
explore_calibration_geo(data, spat_variables, plot = TRUE)
}
\arguments{
\item{data}{an object of class \code{prepare_data} returned by the
\code{\link{prepare_data}}() function.}

\item{spat_variables}{(SpatRaster) predictor variables used for model calibration.}

\item{plot}{(logical) wheter to plot the SpatRaster. Default is TRUE.}
}
\value{
A categorical \code{SpatRaster} with four factors representing:
\describe{
\item{1 - Background cells}{}
\item{2 - Presence cells}{}
\item{3 - Cells with both presence and background}{}
\item{4 - Non-used cells}{}
}
}
\description{
Explore the spatial distribution of occurrence and background points
}
\examples{
#Import raster layers
var <- terra::rast(system.file("extdata", "Current_variables.tif",
                               package = "kuenm2"))
#Import occurrences
data(occ_data, package = "kuenm2")
#Prepare data
sp_swd <- prepare_data(model_type = "glmnet", occ = occ_data,
                       species = occ_data[1, 1], x = "x", y = "y",
                       spat_variables = var, mask = NULL,
                       categorical_variables = "SoilType",
                       do_pca = FALSE, deviance_explained = 95,
                       min_explained = 5, center = TRUE, scale = TRUE,
                       write_pca = FALSE, output_pca = NULL, nbg = 500,
                       kfolds = 4, weights = NULL, min_number = 2,
                       min_continuous = NULL,
                       features = c("l", "q", "p", "lq", "lqp"),
                       regm = c(0.1, 1, 2, 3, 5),
                       include_xy = TRUE,
                       write_file = FALSE, file_name = NULL,
                       seed = 1)
#Explore calibration data
pbg <- explore_calibration_geo(data = sp_swd, spat_variables = var[[1]],
                               plot = FALSE)

}
