% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_calibration_hist.R
\name{explore_calibration_hist}
\alias{explore_calibration_hist}
\title{Explore variable distribution for occurrence and background points}
\usage{
explore_calibration_hist(data, include_m = FALSE, raster_variables = NULL,
                         magnify_occurrences = 2, breaks = 15)
}
\arguments{
\item{data}{an object of class \code{prepared_data} returned by the
\code{\link{prepare_data}}() function.}

\item{include_m}{(logical) whether to include data for plotting the histogram
of the entire area from which background points were sampled. Default is
FALSE, meaning only background and presence information will be plotted.}

\item{raster_variables}{(SpatRaster) predictor variables used to prepared the
data with \code{prepared_data}. Only applicable if \code{include_m} is TRUE.}

\item{magnify_occurrences}{(numeric) factor by which the frequency of
occurrences is magnified for better visualization. Default is 2, meaning
occurrence frequencies in the plot will be doubled.}

\item{breaks}{(numeric) a single number giving the desired number of
intervals in the histogram.}
}
\value{
A list of with information to plot informative histograms to explore data
to be used in the modeling process. Histogram plots can be plotted with
the function \code{\link[=plot_calibration_hist]{plot_calibration_hist()}}.
}
\description{
This function prepares data to generate overlaid histograms to visualize the
distribution of predictor variables for occurrence (presence) and background
points.
}
\examples{
# Import raster layers
var <- terra::rast(system.file("extdata", "Current_variables.tif",
                               package = "kuenm2"))

# Import occurrences
data(sp_swd, package = "kuenm2")

# Explore calibration data
calib_hist <- explore_calibration_hist(data = sp_swd,
                                       raster_variables = var,
                                       include_m = TRUE)

# To visualize results use the function plot_calibration_hist()
}
\seealso{
\code{\link[=plot_calibration_hist]{plot_calibration_hist()}}
}
