% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_var_from_formulas.R
\name{extract_var_from_formulas}
\alias{extract_var_from_formulas}
\title{Extract Predictor Names from Formulas}
\usage{
extract_var_from_formulas(formulas, ...)
}
\arguments{
\item{formulas}{(character or formula) model formulas.}

\item{...}{Arguments to pass to \code{\link[=all.vars]{all.vars()}}}
}
\value{
A character vector or a list of the same length as the formulas,
containing the variable names.
}
\description{
Extract Predictor Names from Formulas
}
\examples{
# Extract variables from the formula grid
# Import an example of calibration results (output from the calibration function)
data("calib_results_glmnet", package = "kuenm2")
extract_var_from_formulas(formulas = calib_results_glmnet$formula_grid$Formulas)

}
