% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_selected.R
\name{fit_selected}
\alias{fit_selected}
\title{Fit models selected after calibration}
\usage{
fit_selected(calibration_results, replicate_method = "kfolds",
             n_replicates = 1, sample_proportion = 0.7, type = "cloglog",
             write_models = FALSE,
             file_name = NULL, parallel = FALSE, ncores = NULL,
             progress_bar = TRUE, verbose = TRUE, seed = 1)
}
\arguments{
\item{calibration_results}{an object of class \code{calibration_results} returned
by the \code{\link[=calibration]{calibration()}} function.}

\item{replicate_method}{(character) method used for producing replicates.
Available options are \code{"kfolds"}, \code{"subsample"}, and \code{"bootstrap"}.
See \strong{Details} for more information.}

\item{n_replicates}{(numeric) number of replicates or folds to generate. If
\code{replicate_method} is \code{"subsample"} or \code{"bootstrap"}, this defines the number
of replicates. If \code{"kfolds"}, it specifies the number of folds. Default is 4.}

\item{sample_proportion}{(numeric) proportion of occurrence and background
points to be used to fit model replicates. Only applicable when
\code{replicate_method} is \code{"subsample"} or \code{"bootstrap"}. Default is 0.7 (i.e.,
70\% of the data).}

\item{type}{(character) the format of prediction values for computing
thresholds. For maxnet models, valid options are "raw", "cumulative",
"logistic", and "cloglog". For glm models, valid options are "cloglog",
"response" and "raw". Default is "cloglog".}

\item{write_models}{(logical) whether to save the final fitted models to disk.
Default is FALSE.}

\item{file_name}{(character) the file name, with or without a path, for saving
the final models. This is only applicable if \code{write_models = TRUE}.}

\item{parallel}{(logical) whether to fit the final models in parallel. Default
is FALSE.}

\item{ncores}{(numeric) number of cores to use for parallel processing.
Default is NULL and uses available cores - 1. This is only applicable if
\code{parallel = TRUE}.}

\item{progress_bar}{(logical) whether to display a progress bar during processing.
Default is TRUE.}

\item{verbose}{(logical) whether to display detailed messages during processing.
Default is TRUE.}

\item{seed}{(numeric) integer value used to specify an initial seed to split
the data. Default is 1.}
}
\value{
An object of class 'fitted_models' containing the following elements:
\item{species}{a character string with the name of the species.}
\item{Models}{a list of fitted models, including replicates (fitted with
part of the data) and full models (fitted with all data).}
\item{calibration_data}{a data.frame containing a column (\code{pr_bg}) that
identifies occurrence points (1) and background points (0), along with the
corresponding values of predictor variables for each point.}
\item{selected_models}{a data frame with the ID and summary of evaluation
metrics for the selected models.}
\item{weights}{a numeric vector specifying weights for the predictor variables
(if used).}
\item{pca}{a list of class \code{\link[terra]{prcomp}} representing the
result of principal component analysis (if performed).}
\item{addsamplestobackground}{a logical value indicating whether any presence
sample not already in the background was added.}
\item{omission_rate}{the omission rate determined during the calibration step.}
\item{thresholds}{the thresholds to binarize each replicate and the consensus
(mean and median), calculated based on the omission rate set in
\code{\link[=calibration]{calibration()}}.}
}
\description{
This function fits models selected during model \code{\link[=calibration]{calibration()}}.
}
\details{
This function also computes model consensus (mean and median), the thresholds
to binarize model predictions based on the omission rate set during model
calibration to select models.
}
\examples{
# An example with maxnet models
data(calib_results_maxnet, package = "kuenm2")

# Fit models using calibration results
fm <- fit_selected(calibration_results = calib_results_maxnet,
                   n_replicates = 4)

# Output the fitted models
fm

# An example with GLMs
data(calib_results_glm, package = "kuenm2")

# Fit models using calibration results
fm_glm <- fit_selected(calibration_results = calib_results_glm,
                       replicate_method = "subsample",
                       n_replicates = 5)

# Output the fitted models
fm_glm
}
