% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_selected_glmnetmx.R
\name{fit_selected_glmnetmx}
\alias{fit_selected_glmnetmx}
\title{Fit selected Maxent-like glmnet models}
\usage{
fit_selected_glmnetmx(
  calibration_results,
  n_replicates = 5,
  rep_type = "kfold",
  train_portion = 0.7,
  write_models = FALSE,
  file_name = NULL,
  parallel = TRUE,
  ncores = 4,
  parallelType = "doSNOW",
  progress_bar = TRUE,
  verbose = TRUE
)
}
\description{
Fit selected Maxent-like glmnet models
}
