% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet_mx.R
\name{glmnet_mx}
\alias{glmnet_mx}
\title{Maxent-like glmnet models}
\usage{
glmnet_mx(p, data, f, regmult = 1.0, regfun = maxnet.default.regularization,
          addsamplestobackground = TRUE, weights = NULL, ...)
}
\arguments{
\item{p}{A vector of binary presence-background labels, where 1 indicates
presence and 0 indicates background.}

\item{data}{A \code{data.frame} containing the predictor variables for the
model. This must include the same number of rows as the length of \code{p}.}

\item{f}{A formula specifying the model to be fitted, in the format used by
\code{\link[stats]{model.matrix}}.}

\item{regmult}{(numeric) Regularization multiplier, default is 1.0.}

\item{regfun}{A function that calculates regularization penalties. Default is
\code{maxnet.default.regularization}.}

\item{addsamplestobackground}{(logical) Whether to add presence points not in
the background to the background data. Default is \code{TRUE}.}

\item{weights}{(numeric) A numeric vector of weights for each observation.
Default is \code{NULL}, which sets weights to 1 for presence points
and 100 for background points.}

\item{...}{Additional arguments to pass to \code{\link[glmnet]{glmnet}}.}
}
\value{
A fitted Maxent-like model object of class \code{glmnet_mx}, which
includes model coefficients, AIC (if requested), and other elements
such as feature mins and maxes, sample means, and entropy.
}
\description{
This function fits Maxent-like models using the \code{glmnet} package, designed
for presence-background data.
}
\details{
This function is modified from the package maxnet and fits a Maxent-like
model using regularization to avoid over-fitting. Regularization weights
are computed using a provided function (which can be changed) and can be
multiplied by a regularization multiplier (\code{regmult}). The function also
includes an option to calculate AIC.
}
