% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuenm2.R
\docType{package}
\name{kuenm2-package}
\alias{kuenm2}
\alias{kuenm2-package}
\title{kuenm2: Detailed Development of Ecological Niche Models}
\description{
kuenm2 A new set of tools to help with the development of detailed ecological
niche models using multiple algorithms. Pre-modeling analyses and
explorations can be done to prepare data. Model calibration (model selection)
can be done by creating and testing several candidate models. Handy options
for producing final models with transfers are included. Other tools to
assess extrapolation risks and variability in model transfers are also
available.
}
\section{Functions by stage in the ENM process}{

\itemize{
\item Data preparation: \code{\link[=initial_cleaning]{initial_cleaning()}}, \code{\link[=advanced_cleaning]{advanced_cleaning()}},
\code{\link[=prepare_data]{prepare_data()}}, \code{\link[=prepare_user_data]{prepare_user_data()}}
\item Model calibration: \code{\link[=calibration]{calibration()}}, \code{\link[=select_models]{select_models()}}
\item Model exploration: \code{\link[=fit_selected]{fit_selected()}}, \code{\link[=variable_importance]{variable_importance()}},
\code{\link[=plot_importance]{plot_importance()}}, \code{\link[=response_curve]{response_curve()}}, \code{\link[=bivariate_response]{bivariate_response()}}
\item Model projection: \code{\link[=predict_selected]{predict_selected()}}, \code{\link[=project_selected]{project_selected()}}
\item Post-modeling: \code{\link[=projection_variability]{projection_variability()}}, \code{\link[=projection_changes]{projection_changes()}},
\code{\link[=projection_mop]{projection_mop()}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://marlonecobos.github.io/kuenm2/}
}

}
\author{
\strong{Maintainer}: Weverton C. F. Trindade \email{wevertonf1993@gmail.com} (\href{https://orcid.org/0000-0003-2045-4555}{ORCID})

Authors:
\itemize{
  \item Luis F. Arias-Giraldo \email{lfarias.giraldo@gmail.com} (\href{https://orcid.org/0000-0003-4861-8064}{ORCID})
  \item Luis Osorio-Olvera \email{luismurao@gmail.com} (\href{https://orcid.org/0000-0003-0701-5398}{ORCID})
  \item A. Townsend Peterson \email{town@ku.edu} (\href{https://orcid.org/0000-0003-0243-2379}{ORCID})
  \item Marlon E. Cobos \email{manubio13@gmail.com} (\href{https://orcid.org/0000-0002-2611-1767}{ORCID})
}

}
