% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_importance.R
\name{plot_importance}
\alias{plot_importance}
\title{Summary plot for variable importance in models}
\usage{
plot_importance(x, xlab = NULL, ylab = "Relative contribution",
                main = "Variable importance", extra_info = TRUE, ...)
}
\description{
See details in \code{\link[enmpa]{plot_importance}}
}
