% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data for model calibration}
\usage{
prepare_data(algorithm, occ, x, y, raster_variables, species = NULL,
             mask = NULL, n_background = 1000, features = c("lq", "lqp"),
             r_multiplier = c(0.1, 0.5, 1, 2, 3), kfolds = 4,
             categorical_variables = NULL, do_pca = FALSE, center = TRUE,
             scale = TRUE, exclude_from_pca = NULL, variance_explained = 95,
             min_explained = 5, min_number = 2, min_continuous = NULL,
             bias_file = NULL, bias_effect = NULL, weights = NULL,
             include_xy = TRUE, write_pca = FALSE, pca_directory = NULL,
             write_file = FALSE, file_name = NULL, seed = 1)
}
\arguments{
\item{algorithm}{(character) modeling algorithm, either "glm" or "maxnet".}

\item{occ}{(data frame) a data.frame containing the coordinates (longitude
and latitude) of the occurrence records.}

\item{x}{(character) a string specifying the name of the column in \code{occ} that
contains the longitude values.}

\item{y}{(character) a string specifying the name of the column in \code{occ} that
contains the latitude values.}

\item{raster_variables}{(SpatRaster) predictor variables used to calibrate the
models.}

\item{species}{(character) string specifying the species name (optional).
Default is NULL.}

\item{mask}{spatial object used to mask the variables to the area where the
model will be calibrated. Mask must be of one of the following classes:
\code{SpatRaster}, \code{SpatVector}, or \code{SpatExtent}. Default is NULL.}

\item{n_background}{(numeric) number of points to represent the background
for the model. Default is 1000.}

\item{features}{(character) a vector of feature classes. Default is c("q",
"lq", "lp", "qp", "lqp").}

\item{r_multiplier}{(numeric) a vector of regularization parameters for maxnet.
Default is c(0.1, 1, 2, 3, 5).}

\item{kfolds}{(numeric) the number of groups (folds) the occurrence
data will be split into for cross-validation. Default is 4.}

\item{categorical_variables}{(character) names of the variables that are
categorical. Default is NULL.}

\item{do_pca}{(logical) whether to perform a principal component analysis
(PCA) with the set of variables. Default is FALSE.}

\item{center}{(logical) whether the variables should be zero-centered. Default
is TRUE.}

\item{scale}{(logical) whether the variables should be scaled to have unit
variance before the analysis takes place. Default is FALSE.}

\item{exclude_from_pca}{(character) variable names within raster_variables
that should not be included in the PCA transformation. Instead, these
variables will be added directly to the final set of output variables
without being modified. The default is NULL, meaning all variables
will be used unless specified otherwise.}

\item{variance_explained}{(numeric) the cumulative percentage of total
variance that must be explained by the selected principal components.
Default is 95.}

\item{min_explained}{(numeric) the minimum percentage of total variance that
a principal component must explain to be retained. Default is 5.}

\item{min_number}{(numeric) the minimum number of variables to be included in
the model formulas to be generated.}

\item{min_continuous}{(numeric) the minimum number of continuous variables
required in a combination. Default is NULL.}

\item{bias_file}{(SpatRaster) a raster containing bias values (probability
weights) that influence the selection of background points. It must have the
same extent, resolution, and number of cells as the raster variables, unless
a mask is provided. Default is NULL.}

\item{bias_effect}{(character) a string specifying how the values in the
\code{bias_file} should be interpreted. Options are "direct" or "inverse". If
"direct", higher values in bias file increase the likelihood of selecting
background points. If "inverse", higher values decrease the likelihood.
Default = NULL. Must be defined if \code{bias_file} is provided.}

\item{weights}{(numeric) a numeric vector specifying weights for the
occurrence records. Default is NULL.}

\item{include_xy}{(logical) whether to include the coordinates (longitude and
latitude) in the results from preparing data. Columns containing coordinates
will be renamed as "x" and "y". Default is TRUE.}

\item{write_pca}{(logical) whether to save the PCA-derived raster layers
(principal components) to disk. Default is FALSE.}

\item{pca_directory}{(character) the path or name of the folder where the PC
raster layers will be saved. This is only applicable if \code{write_pca = TRUE}.
Default is NULL.}

\item{write_file}{(logical) whether to write the resulting prepared_data list
in a local directory. Default is FALSE.}

\item{file_name}{(character) the path or name of the folder where the
resulting list will be saved. This is only applicable if \code{write_file = TRUE}. Default is NULL.}

\item{seed}{(numeric) integer value to specify an initial seed to split the
data and extract background. Default is 1.}
}
\value{
An object of class \code{prepared_data} containing all elements to run a model
calibration routine. The elements include: species, calibration data,
a grid of model parameters, indices of k-folds for cross validation,
xy coordinates, names of continuous and categorical variables, weights,
results from PCA, and modeling algorithm.
}
\description{
This function prepares data for model calibration, including optional PCA,
background point generation, k-fold partitioning, and the creation of a grid
of parameter combinations, including regularization multiplier values,
feature classes, and sets of environmental variables.
}
\examples{
# Import occurrences
data(occ_data, package = "kuenm2")

# Import raster layers
var <- terra::rast(system.file("extdata", "Current_variables.tif",
                               package = "kuenm2"))

# Import a bias file
bias <- terra::rast(system.file("extdata", "bias_file.tif",
                                package = "kuenm2"))

# Prepare data for maxnet model
sp_swd <- prepare_data(algorithm = "maxnet", occ = occ_data,
                       x = "x", y = "y",
                       raster_variables = var,
                       species = occ_data[1, 1],
                       categorical_variables = "SoilType",
                       n_background = 500, bias_file = bias,
                       bias_effect = "direct",
                       features = c("l", "q", "p", "lq", "lqp"),
                       r_multiplier = c(0.1, 1, 2, 3, 5))
print(sp_swd)

# Prepare data for glm model
sp_swd_glm <- prepare_data(algorithm = "glm", occ = occ_data,
                           x = "x", y = "y",
                           raster_variables = var,
                           species = occ_data[1, 1],
                           categorical_variables = "SoilType",
                           n_background = 500, bias_file = bias,
                           bias_effect = "direct",
                           features = c("l", "q", "p", "lq", "lqp"))
print(sp_swd_glm)
}
