% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_proj.R
\name{prepare_proj}
\alias{prepare_proj}
\title{Preparation of data for model projections}
\usage{
prepare_proj(
  models = NULL,
  variable_names = NULL,
  present_dir = NULL,
  past_dir = NULL,
  past_time = NULL,
  past_gcm = NULL,
  future_dir = NULL,
  future_time = NULL,
  future_pscen = NULL,
  future_gcm = NULL,
  write_file = FALSE,
  filename = NULL,
  raster_pattern = ".tif*"
)
}
\description{
Preparation of data for model projections
}
