% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res2var.R
\name{resp2var}
\alias{resp2var}
\title{Two-Way interaction response plot}
\usage{
resp2var(models, modelID, variable1 , variable2, n = 1000,
         new_data = NULL, extrapolate = FALSE, add_bar = TRUE ,
         add_limits = FALSE, color.palette	= NULL,
         xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{models}{an object of class \code{fitted_models} returned by the
\code{\link{fit_selected}}() function.}

\item{modelID}{(character) name of the ModelID presents in the fitted
object. Default = NULL.}

\item{variable1}{(character) name of the variable to be plotted in x axis.}

\item{variable2}{(character) name of the variable to be plotted in y axis.}

\item{n}{(numeric) an integer guiding the number of breaks. Default = 1000}

\item{new_data}{a \code{SpatRaster}, data.frame, or  matrix of variables
representing the range of variable values in an area of interest.
Default = NULL.}

\item{extrapolate}{(logical) whether to allow extrapolation to study the
behavior of the response outside the calibration limits. Ignored if
\code{new_data} is defined. Default = TRUE.}

\item{add_bar}{(logical) whether to add bar legend. Default = TRUE.}

\item{add_limits}{(logical) whether to add calibration limits if
\code{extrapolate = TRUE}. Default = FALSE.}

\item{color.palette}{(function) a color palette function to be used to assign
colors in the plot. Default = function(n) rev(hcl.colors(n, "terrain")).}

\item{xlab}{(character) a label for the x axis. The default, NULL, uses the
name defined in \code{variable1}.}

\item{ylab}{(character) a label for the y axis. The default, NULL, uses the
name defined in \code{variable2}.}

\item{...}{additional arguments passed to
\code{\link[graphics]{image}}.}
}
\value{
A plot with the response interaction of two environmental dimensions for
\code{variable1} and \code{variable2}, and don't return anything.
}
\description{
A view of the species probability into a two-dimensional
environmental space.
}
\examples{
##Example with glmnet
# Import example of fitted_models (output of fit_selected())
data("fitted_model_glmnet", package = "kuenm2")

#Response curves (products)
resp2var(models = fitted_model_glmnet, modelID = "Model_1",
         variable1 = "bio_1", variable2 = "bio_7")
resp2var(models = fitted_model_glmnet, modelID = "Model_13",
         variable1 = "bio_1", variable2 = "bio_12")
resp2var(models = fitted_model_glmnet, modelID = "Model_13",
         variable1 = "bio_7", variable2 = "bio_12")

##Example with glm
# Import example of fitted_models (output of fit_selected())
data("fitted_model_glm", package = "kuenm2")

#Response curves
resp2var(models = fitted_model_glm, modelID = "Model_4",
         variable1 = "bio_1", variable2 = "bio_12")
resp2var(models = fitted_model_glm, modelID = "Model_4",
         variable1 = "bio_1", variable2 = "bio_7")
resp2var(models = fitted_model_glm, modelID = "Model_4",
         variable1 = "bio_7", variable2 = "bio_12")

}
