% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res2var.R
\name{resp2var}
\alias{resp2var}
\title{Two-Way interaction response plot}
\usage{
resp2var(fitted, modelID, variable1 , variable2, n = 100,
         new_data = NULL, extrapolate = FALSE, add_bar = TRUE ,
         add_limits = FALSE, color.palette	= NULL,
         xlab = NULL, ylab = NULL, ...)
}
\value{
A plot with the response interaction of two environmental dimensions for
\code{variable1} and \code{variable2}, and don't return anything.
}
\description{
A view of the species probability into a two-dimensional
environmental space.
}
