% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_curves.R
\name{response_curve}
\alias{response_curve}
\title{Variable response curves}
\usage{
response_curve(models, variable, modelID = NULL, n = 100,
                     by_replicates = FALSE, data = NULL, new_data = NULL,
                     extrapolate = TRUE, xlab = NULL, ylab = "Suitability",
                     col = "darkblue", ...)
}
\arguments{
\item{models}{an object of class \code{fitted_models} returned by the
\code{\link{fit_selected}}() function.}

\item{variable}{(character) name of the variables to be plotted.}

\item{modelID}{(character) vector of ModelID(s) to be considered in the
models object. By default all models are included.Default = NULL.}

\item{n}{(numeric) an integer guiding the number of breaks. Default = 100}

\item{by_replicates}{(logical) whether use replicates or full_model to
estimate the model's response curve. Default = FALSE.}

\item{data}{data.frame or matrix of data used in the model calibration step.
Default = NULL.}

\item{new_data}{a \code{SpatRaster}, data.frame, or  matrix of variables
representing the range of variable values in an area of interest.
Default = NULL. It must be defined in case the model entered does not
explicitly include a data component.}

\item{extrapolate}{(logical) whether to allow extrapolation to study the
behavior of the response outside the calibration limits. Ignored if
\code{new_data} is defined. Default = TRUE.}

\item{xlab}{(character) a label for the x axis. The default, NULL, uses the
name defined in \code{variable}.}

\item{ylab}{(character) a label for the y axis. Default = "Suitability".}

\item{col}{(character) color for lines. Default = "darkblue".}

\item{...}{additional arguments passed to \code{\link[graphics]{plot}}.}
}
\value{
A plot with the response curve for a \code{variable}.
}
\description{
A view of variable responses in fitted models. Responses based on single or multiple
models can be provided.
}
\examples{
##Example with glmnet
# Import example of fitted_models (output of fit_selected())
data("fitted_model_glmnet", package = "kuenm2")

#Response curves
response_curve(models = fitted_model_glmnet,
               variable = "bio_1", by_replicates = T)
response_curve(models = fitted_model_glmnet,
               variable = "bio_7", by_replicates = T)
response_curve(models = fitted_model_glmnet, variable = "bio_1",
               modelID = "Model_13", by_replicates = T)
response_curve(models = fitted_model_glmnet, variable = "bio_1",
               modelID = "Model_1", by_replicates = T)

##Example with glm
# Import example of fitted_models (output of fit_selected())
data("fitted_model_glm", package = "kuenm2")

#Response curves
response_curve(models = fitted_model_glm,
               variable = "bio_1", by_replicates = T)
response_curve(models = fitted_model_glm,
               variable = "bio_7", by_replicates = T)
response_curve(models = fitted_model_glm, variable = "bio_1",
               modelID = "Model_4", by_replicates = T)
response_curve(models = fitted_model_glm, variable = "bio_12",
               modelID = "Model_4", by_replicates = T)

}
