% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_curves.R
\name{response_curve}
\alias{response_curve}
\title{Generate Variable Response Curves from Fitted Models}
\usage{
response_curve(models, variable, modelID = NULL, n = 100,
                     by_replicates = FALSE, data = NULL, new_data = NULL,
                     averages_from = "pr",
                     extrapolate = TRUE, extrapolation_factor = 0.1,
                     l_limit = NULL, u_limit = NULL,
                     xlab = NULL, ylab = "Suitability",
                     col = "darkblue", ...)
}
\arguments{
\item{models}{an object of class \code{fitted_models} returned by the
\code{\link{fit_selected}}() function.}

\item{variable}{(character) name of the variable to be plotted.}

\item{modelID}{(character) vector of ModelID(s) to be considered in the
models object. By default all models are included.Default = NULL.}

\item{n}{(numeric) an integer guiding the number of breaks. Default = 100}

\item{by_replicates}{(logical) whether use replicates or full_model to
estimate the model's response curve. Default = FALSE.}

\item{data}{data.frame or matrix of data used in the model calibration step.
Default = NULL.}

\item{new_data}{a \code{SpatRaster}, data.frame, or  matrix of variables
representing the range of variable values in an area of interest.
Default = NULL. It must be defined in case the model entered does not
explicitly include a data component.}

\item{averages_from}{(character) specifies how the averages or modes of the
variables are calculated. Available options are "pr" (to calculate averages
from the presence localities) or "pr_bg" (to use the combined set of presence
and background localities). Default is "pr". See details.}

\item{extrapolate}{(logical) whether to allow extrapolation to study the
behavior of the response outside the calibration limits. Ignored if
\code{new_data} is defined. Default = TRUE.}

\item{extrapolation_factor}{(numeric) a multiplier used to calculate the
extrapolation range. Larger values allow broader extrapolation beyond the
observed data range. Default is 0.1.}

\item{l_limit}{(numeric) specifies the lower limit for the variable. Default
is \code{NULL}, meaning the lower limit will be calculated based on the
data's minimum value and the \code{extrapolation_factor}
(if \code{extrapolation = TRUE}).}

\item{u_limit}{(numeric) specifies the upper limit for the variable. Default
is \code{NULL}, meaning the upper limit will be calculated based on the
data's minimum value and the \code{extrapolation_factor}
(if \code{extrapolation = TRUE}).}

\item{xlab}{(character) a label for the x axis. The default, NULL, uses the
name defined in \code{variable}.}

\item{ylab}{(character) a label for the y axis. Default = "Suitability".}

\item{col}{(character) color for lines. Default = "darkblue".}

\item{...}{additional arguments passed to \code{\link[graphics]{plot}}.}
}
\value{
A plot with the response curve for a \code{variable}.
}
\description{
A view of variable responses in fitted models. Responses based on single or multiple
models can be provided.
}
\details{
The response curves are generated with all other variables set to their mean
values (or mode for categorical variables), calculated either from the
presence localities (if averages_from = "pr") or from the combined set of
presence and background localities (if averages_from = "pr_bg").

For categorical variables, a bar plot is generated with error bars showing
variability across models (if multiple models are included).
}
\examples{
##Example with glmnet
# Import example of fitted_models (output of fit_selected())
data("fitted_model_glmnet", package = "kuenm2")

#Response curves
response_curve(models = fitted_model_glmnet,
               variable = "bio_1", by_replicates = T)
response_curve(models = fitted_model_glmnet,
               variable = "bio_7", by_replicates = T)
response_curve(models = fitted_model_glmnet, variable = "bio_1",
               modelID = "Model_13", by_replicates = T)
response_curve(models = fitted_model_glmnet, variable = "bio_1",
               modelID = "Model_1", by_replicates = T)

##Example with glm
# Import example of fitted_models (output of fit_selected())
data("fitted_model_glm", package = "kuenm2")

#Response curves
response_curve(models = fitted_model_glm,
               variable = "bio_1", by_replicates = T)
response_curve(models = fitted_model_glm,
               variable = "bio_7", by_replicates = T)
response_curve(models = fitted_model_glm, variable = "bio_1",
               modelID = "Model_4", by_replicates = T)
response_curve(models = fitted_model_glm, variable = "bio_12",
               modelID = "Model_4", by_replicates = T)

}
