% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_curves.R
\name{response_curve}
\alias{response_curve}
\title{Variable response curves}
\usage{
response_curve(fitted, variable, modelID = NULL, n = 100,
               by_replicates = FALSE, data = NULL, new_data = NULL,
               extrapolate = TRUE, xlab = NULL, ylab = "Suitability",
               col = "darkblue" ...)
}
\arguments{
\item{fitted}{object.}

\item{variable}{(character) name of the variables to be plotted.}

\item{modelID}{(character) vector of ModelID(s) to be considered in the
fitted object. By default all models are included.Default = NULL.}

\item{n}{(numeric) an integer guiding the number of breaks. Default = 100}

\item{by_replicates}{(logical) whether use replicates or full_model to
estimate the model's response curve.}

\item{data}{data.frame or matrix of data used in the model calibration step.
Default = NULL.}

\item{new_data}{a \code{SpatRaster}, data.frame, or  matrix of variables
representing the range of variable values in an area of interest.
Default = NULL. It must be defined in case the model entered does not
explicitly include a data component.}

\item{extrapolate}{(logical) whether to allow extrapolation to study the
behavior of the response outside the calibration limits. Ignored if
\code{new_data} is defined. Default = TRUE.}

\item{xlab}{(character) a label for the x axis. The default, NULL, uses the
name defined in \code{variable}.}

\item{ylab}{(character) a label for the y axis. Default = "Suitability".}

\item{col}{(character) color for lines. Default = "darkblue".}

\item{...}{additional arguments passed to \code{\link[graphics]{plot}}.}
}
\value{
A plot with the response curve for a \code{variable}.
}
\description{
A view of variable responses in models. Responses based on single or multiple
models can be provided.
}
