% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_importance.R
\name{var_importance}
\alias{var_importance}
\title{Variable importance}
\usage{
var_importance(fitted, modelID = NULL)
}
\value{
A data.frame containing the relative contribution of each variable. An
identification for distinct models is added if \code{fitted} contains multiple
models.
}
\description{
Variable importance
}
